rename task, `eodms_prompt` and `eodms_cli`
