+++ .github/workflows/cli_tests.yml
@@ -38,49 +38,49 @@ jobs:
 
       # Set 'download_attempts' in configuration file
       - name: Set 'download_attempts' in configuration file
-        run: python eodms_cli.py --configure RAPI.download_attempts=20
+        run: python eodms_prompt.py --configure RAPI.download_attempts=20
 
       # Runs a single command using the runners shell
       - name: Run EODMS-CLI - Process 1
         env:
           EODMS_USER: ${{ secrets.EODMS_USER }}
           EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-        run: python eodms_cli.py -u $EODMS_USER -p $EODMS_PASSWORD --collections RCMImageProducts,Radarsat2 -d 20190101-20220527 -i test/files/NCR_AOI.geojson -max 2 -prc full -ov 30 -f "RCMImageProducts.beam_mnemonic like 16M%%,RCMImageProducts.product_type=SLC,Radarsat2.beam_mnemonic like EH%%,Radarsat2.transmit_polarization=H" -o test/files/test1_auto.geojson -dn test/downloads -s
+        run: python eodms_prompt.py -u $EODMS_USER -p $EODMS_PASSWORD --collections RCMImageProducts,Radarsat2 -d 20190101-20220527 -i test/files/NCR_AOI.geojson -max 2 -prc full -ov 30 -f "RCMImageProducts.beam_mnemonic like 16M%%,RCMImageProducts.product_type=SLC,Radarsat2.beam_mnemonic like EH%%,Radarsat2.transmit_polarization=H" -o test/files/test1_auto.geojson -dn test/downloads -s
 
       # Runs a single command using the runners shell
       - name: Run EODMS-CLI - Process 2
         env:
           EODMS_USER: ${{ secrets.EODMS_USER }}
           EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-        run: python eodms_cli.py -u $EODMS_USER -p $EODMS_PASSWORD -i test/files/RCMImageProducts_Results.csv -max 4 -prc order_csv -o test/files/test2_auto.geojson -s
+        run: python eodms_prompt.py -u $EODMS_USER -p $EODMS_PASSWORD -i test/files/RCMImageProducts_Results.csv -max 4 -prc order_csv -o test/files/test2_auto.geojson -s
 
       # Runs a single command using the runners shell
       - name: Run EODMS-CLI - Process 3
         env:
           EODMS_USER: ${{ secrets.EODMS_USER }}
           EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-        run: python eodms_cli.py -u $EODMS_USER -p $EODMS_PASSWORD --collection RCMImageProducts --uuid "13531983,13531917" -prc uuid -o test/files/test3_auto.geojson -s
+        run: python eodms_prompt.py -u $EODMS_USER -p $EODMS_PASSWORD --collection RCMImageProducts --uuid "13531983,13531917" -prc uuid -o test/files/test3_auto.geojson -s
 
       # Runs a single command using the runners shell
       - name: Run EODMS-CLI - Process 4
         env:
           EODMS_USER: ${{ secrets.EODMS_USER }}
           EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-        run: python eodms_cli.py -u $EODMS_USER -p $EODMS_PASSWORD -prc download_restored_items -i test/files/20260216_155955_ItemsRestoring.json -s
+        run: python eodms_prompt.py -u $EODMS_USER -p $EODMS_PASSWORD -prc download_restored_items -i test/files/20260216_155955_ItemsRestoring.json -s
 
       # Runs a single command using the runners shell
       - name: Run EODMS-CLI - Process 5
         env:
           EODMS_USER: ${{ secrets.EODMS_USER }}
           EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-        run: python eodms_cli.py -u $EODMS_USER -p $EODMS_PASSWORD -prc order_st -st files/sar_toolbox_request.json -o test/files/test5_output.geojson -s
+        run: python eodms_prompt.py -u $EODMS_USER -p $EODMS_PASSWORD -prc order_st -st files/sar_toolbox_request.json -o test/files/test5_output.geojson -s
 
       # Runs a single command using the runners shell
       # - name: Run EODMS-CLI - Process 6
       #   env:
       #     EODMS_USER: ${{ secrets.EODMS_USER }}
       #     EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-      #   run: python eodms_cli.py -u $EODMS_USER -p $EODMS_PASSWORD -prc order_st -st test/files/sar_toolbox_request.json -o test/test6_auto.geojson -s
+      #   run: python eodms_prompt.py -u $EODMS_USER -p $EODMS_PASSWORD -prc order_st -st test/files/sar_toolbox_request.json -o test/test6_auto.geojson -s
 
       # List files
       - name: List files
