+++ .github/workflows/prompt_tests.yml
@@ -38,7 +38,7 @@ jobs:
         
       # Set 'download_attempts' in configuration file
       - name: Set 'download_attempts' in configuration file
-        run: python eodms_cli.py --configure RAPI.download_attempts=20
+        run: python eodms_prompt.py --configure RAPI.download_attempts=20
 
       # Runs a single command using the runners shell
       - name: Run EODMS-CLI Prompt Test 1 - Process 1
@@ -46,7 +46,7 @@ jobs:
         env:
           EODMS_USER: ${{ secrets.EODMS_USER }}
           EODMS_PASSWORD: ${{ secrets.EODMS_PWD }}
-        # run: python eodms_cli.py -u $USERNAME -p $PASSWORD --collections RCMImageProducts,Radarsat2 -d 20190101-20220527 -i tests/files/NCR_AOI.geojson -max 2:1 -prc full -ov 30 -f "RCMImageProducts.beam_mnemonic like 16M%%,RCMImageProducts.product_type=SLC,Radarsat2.beam_mnemonic like EH%%,Radarsat2.transmit_polarization=H" -o tests/files/test1_output.geojson -pri low -s
+        # run: python eodms_prompt.py -u $USERNAME -p $PASSWORD --collections RCMImageProducts,Radarsat2 -d 20190101-20220527 -i tests/files/NCR_AOI.geojson -max 2:1 -prc full -ov 30 -f "RCMImageProducts.beam_mnemonic like 16M%%,RCMImageProducts.product_type=SLC,Radarsat2.beam_mnemonic like EH%%,Radarsat2.transmit_polarization=H" -o tests/files/test1_output.geojson -pri low -s
         run: |
           cd test
           python -m unittest test_prompt.TestEodmsCli.test_process1
