+++ eodms_cli.py
@@ -1,2108 +1,1717 @@
-##############################################################################
-#
-# Copyright (c) His Majesty the King in Right of Canada, as
-# represented by the Minister of Natural Resources, 2026
-# 
-# Licensed under the MIT license
-# (see LICENSE or <http://opensource.org/licenses/MIT>) All files in the 
-# project carrying such notice may not be copied, modified, or distributed 
-# except according to those terms.
-# 
-##############################################################################
+"""EODMS CLI v2 (non-interactive) focused on eodms-py STAC/DDS workflows.
 
-__title__ = 'EODMS-CLI'
-__author__ = 'Kevin Ballantyne'
-__copyright__ = 'Copyright (c) His Majesty the King in Right of Canada, ' \
-                'as represented by the Minister of Natural Resources, 2026'
-__license__ = 'MIT License'
-__description__ = 'Script used to search, order and download imagery from ' \
-                  'the EODMS using the REST API (RAPI) service and DDS API service.'
-__version__ = '4.1.2'
-__maintainer__ = 'Kevin Ballantyne'
-__email__ = 'eodms-sgdot@nrcan-rncan.gc.ca'
+This module provides the current command-driven CLI entrypoint with three
+core flows:
+- `search`: STAC catalog discovery, queryables, and GeoJSON export.
+- `process`: OGC Processes list/inspect/submit/job workflows.
+- `download`: STAC asset, DDS UUID, and legacy RAPI order-item downloads.
 
-import sys
-import os
-import re
-import requests
-# from urllib.parse import urlparse
-# import argparse
-import click
-import traceback
-import datetime
-import textwrap
-from typing import List, Dict, Any
-# from geomet import wkt
+Legacy prompt-driven workflows are implemented separately in eodms_prompt.py.
+"""
+
 import json
-# import configparser
+import os
+import configparser
 import base64
 import binascii
-import logging
-import logging.handlers as handlers
-import pathlib
-from colorama import Fore, Back, Style
-# from distutils.version import LooseVersion
-# from distutils.version import StrictVersion
-from packaging import version as pack_v
-# import unicodedata
-import eodms_rapi
+import time
+from functools import wraps
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from datetime import datetime, timedelta, timezone
+from urllib.error import URLError
+from urllib.parse import quote, urlencode, unquote, urlsplit
+from urllib.request import urlopen
+from typing import Any, Dict, List, Optional, Tuple
+
+import click
+import fiona
 from shapely.geometry import shape
-from shapely.ops import unary_union
-from eodms.search import Search_API
 
+from eodms import Processes_API, aaa, dds, search
 try:
-    import fiona
+    from eodms.errors import EODMSError
+    EODMS_SERVICE_ERRORS = (EODMSError,)
 except Exception:
-    fiona = None
+    # Backward-compatible fallback for eodms versions that do not expose EODMSError.
+    from eodms.errors import CatalogError, DDSError, ProcessingError, SearchError
+    EODMS_SERVICE_ERRORS = (CatalogError, DDSError, ProcessingError, SearchError)
+from eodms_rapi import EODMSRAPI, QueryError
 
-# from eodms_rapi import EODMSRAPI
 
-from scripts import utils as eod_util
-from scripts import field
-from scripts import config_util
-from scripts import sar
+DEFAULT_DDS_BACKOFF_SECONDS = 60
+DEFAULT_DDS_CONCURRENT_DOWNLOADS = 10
+DEFAULT_DDS_RETRY_FILE = ".\\downloads\\dds_retry_items.jsonl"
+MAX_DDS_QUEUED_WAITS = 10
 
-# from utils import csv_util
-# from utils import image
-# from utils import geo
 
-proc_choices = {'full': {
-                    'name': 'Search and/or download',
-                    'desc': 'Search and/or download images'
-                },
-                'order_csv': {
-                    'name': 'EODMS UI Ordering',
-                    'desc': 'Download images using EODMS UI search '
-                            'results (CSV file)'
-                },
-                'uuid': {
-                    'name': 'UUID',
-                    'desc': 'Order and download a single or set of images '
-                            'using UUID(s)'
-                },
-                'download_restored_items': {
-                    'name': 'Download Restored Items',
-                    'desc': 'Download requests that had status ItemsRestoring'
-                },
-                'order_st': {
-                    'name': 'Submit Order to SAR Toolbox',
-                    'desc': 'Submit order to the SAR Toolbox'
-                },
-                'download_available': {
-                    'name': 'Download Available SAR Toolbox Order Items',
-                    'desc': 'Downloads Order Items with status '
-                            'AVAILABLE_FOR_DOWNLOAD for SAR Toolbox orders only.'
-                }
-            }
+class OrderedHelpGroup(click.Group):
+    """Group that prints commands in a custom help order."""
 
-min_rapi_version = '1.9.0'
+    def list_commands(self, ctx):
+        preferred = ["search", "process", "download"]
+        remaining = sorted(name for name in self.commands if name not in preferred)
+        return [name for name in preferred if name in self.commands] + remaining
+
+
+class ServiceError(click.ClickException):
+    exit_code = 2
+
+
+def handle_service_errors(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except EODMS_SERVICE_ERRORS as exc:
+            raise ServiceError(f"Service error: {exc}") from exc
+
+    return wrapper
+
+
+def _default_config_path() -> str:
+    user_home = os.environ.get("USERPROFILE") or os.path.expanduser("~")
+    return os.path.join(user_home, ".eodms", "config.ini")
+
+
+def _load_config_parser(config_path: Optional[str] = None) -> Optional[configparser.ConfigParser]:
+    cfg_path = config_path or _default_config_path()
+    if not os.path.exists(cfg_path):
+        return None
+
+    parser = configparser.ConfigParser(comment_prefixes='/', allow_no_value=True)
+    parser.read(cfg_path)
+    return parser
+
+
+def _decode_config_password(raw_password: str) -> str:
+    try:
+        return base64.b64decode(raw_password).decode("utf-8")
+    except binascii.Error:
+        return base64.b64decode(raw_password + "========").decode("utf-8")
+
+
+def _load_credentials_from_config(config_path: Optional[str] = None) -> Tuple[Optional[str], Optional[str]]:
+    parser = _load_config_parser(config_path)
+    if parser is None:
+        return None, None
+
+    username = None
+    password = None
+
+    for section in ("Credentials", "RAPI"):
+        if username is None and parser.has_option(section, "username"):
+            candidate = parser.get(section, "username").strip()
+            if candidate:
+                username = candidate
+
+        if password is None and parser.has_option(section, "password"):
+            encoded = parser.get(section, "password").strip()
+            if encoded:
+                try:
+                    password = _decode_config_password(encoded)
+                except Exception:
+                    password = encoded
+
+    return username, password
+
+
+def resolve_credentials(username: Optional[str], password: Optional[str]) -> Tuple[Optional[str], Optional[str]]:
+    if username and password:
+        return username, password
+
+    cfg_username, cfg_password = _load_credentials_from_config()
+    return username or cfg_username, password or cfg_password
+
+
+def _load_dds_backoff_interval(config_path: Optional[str] = None) -> int:
+    parser = _load_config_parser(config_path)
+    if parser is None or not parser.has_section("DDS"):
+        return DEFAULT_DDS_BACKOFF_SECONDS
+
+    for option_name in ("backoff_interval", "back_interval"):
+        if parser.has_option("DDS", option_name):
+            value = parser.get("DDS", option_name).strip()
+            try:
+                parsed = int(value)
+                if parsed > 0:
+                    return parsed
+            except ValueError:
+                continue
+
+    return DEFAULT_DDS_BACKOFF_SECONDS
+
+
+def _load_dds_concurrent_downloads(config_path: Optional[str] = None) -> int:
+    parser = _load_config_parser(config_path)
+    if parser is None or not parser.has_section("DDS"):
+        return DEFAULT_DDS_CONCURRENT_DOWNLOADS
+
+    if parser.has_option("DDS", "concurrent_downloads"):
+        value = parser.get("DDS", "concurrent_downloads").strip()
+        try:
+            parsed = int(value)
+            if parsed > 0:
+                return parsed
+        except ValueError:
+            pass
+
+    return DEFAULT_DDS_CONCURRENT_DOWNLOADS
+
+
+def _save_credentials_to_config(username: str, password: str,
+                                config_path: Optional[str] = None) -> str:
+    cfg_path = config_path or _default_config_path()
+    cfg_dir = os.path.dirname(cfg_path)
+    if cfg_dir:
+        os.makedirs(cfg_dir, exist_ok=True)
+
+    parser = configparser.ConfigParser(comment_prefixes='/', allow_no_value=True)
+    if os.path.exists(cfg_path):
+        parser.read(cfg_path)
+
+    if not parser.has_section("Credentials"):
+        parser.add_section("Credentials")
+
+    parser.set("Credentials", "username", username)
+    encoded_password = base64.b64encode(password.encode("utf-8")).decode("utf-8")
+    parser.set("Credentials", "password", encoded_password)
+
+    with open(cfg_path, "w", encoding="utf-8") as cfg_file:
+        parser.write(cfg_file, space_around_delimiters=True)
+
+    return cfg_path
+
 
 def parse_aoi_file(aoi_file: str) -> List[Dict[str, Any]]:
-    """
-    Read polygon features from a GeoJSON, shapefile, or geopackage and return WKT strings.
-    
-    :param aoi_file: Path to an AOI geospatial file.
-    :type  aoi_file: str
-    
-    :return: List of dicts with 'name' and 'wkt' keys.
-    :rtype: List[Dict[str, Any]]
-    
-    :raises ValueError: If file cannot be opened or contains no valid polygons.
-    """
+    """Read polygon features from a geospatial file and return WKT geometries."""
     try:
         with fiona.open(aoi_file) as src:
             features = list(src)
-    except Exception as e:
-        raise ValueError(f"Could not open AOI file '{aoi_file}': {e}")
+    except Exception as exc:
+        raise ValueError(f"Could not open AOI file '{aoi_file}': {exc}")
 
-    polygons = []
+    polygons: List[Dict[str, Any]] = []
     for feature in features:
-        geom = feature.get('geometry')
-        if geom is None or geom.get('type') not in ('Polygon', 'MultiPolygon'):
+        geom = feature.get("geometry")
+        if geom is None or geom.get("type") not in ("Polygon", "MultiPolygon"):
             continue
-        name = (feature.get('properties') or {}).get('name')
-        polygons.append({'name': name, 'wkt': shape(geom).wkt})
+        name = (feature.get("properties") or {}).get("name")
+        polygons.append({"name": name, "wkt": shape(geom).wkt})
 
     if not polygons:
         raise ValueError("No polygon geometries found in AOI file.")
     if len(polygons) > 5:
         raise ValueError(f"AOI file contains {len(polygons)} polygons; maximum is 5.")
 
-    print(f"Loaded {len(polygons)} polygon(s) from AOI file.")
     return polygons
 
-class Prompter:
-    """
-    Class used to prompt the user for all inputs.
-    """
 
-    def __init__(self, eod, config_util, params, testing=False):
-        """
-        Initializer for the Prompter class.
-        
-        :param eod: The Eodms_OrderDownload object.
-        :type  eod: self.Eodms_OrderDownload
-        :param config_util: The ConfigUtils object
-        :type  config_util: ConfigUtils
-        :param params: An empty dictionary of parameters.
-        :type  params: dict
-        """
+def parse_bbox(bbox: Optional[str]) -> Optional[List[float]]:
+    if not bbox:
+        return None
 
-        self.eod = eod
-        self.eod.set_prompter(self)
-        self.reset_col = eod.get_colour(reset=True)
-        self.config_util = config_util
-        self.config_info = config_util.get_info()
-        self.params = params
-        self.process = None
-        self.testing = testing
+    try:
+        bbox_list = [float(x.strip()) for x in bbox.split(",")]
+    except ValueError as exc:
+        raise click.BadParameter(f"Invalid bbox values: {exc}")
 
-        self.logger = logging.getLogger('eodms')
+    if len(bbox_list) != 4:
+        raise click.BadParameter("Bounding box must contain exactly 4 values: west,south,east,north")
 
-    def _parse_aoi_file(self, input_fn):
-        """
-        Parse an AOI geospatial file using Fiona and return WKT.
+    return bbox_list
 
-        :param input_fn: Path to an AOI file.
-        :type  input_fn: str
 
-        :return: WKT geometry string or None if invalid.
-        :rtype: str or None
-        """
+def save_items_geojson(items: Optional[List[Dict[str, Any]]], output_file: str) -> None:
+    feature_collection = {
+        "type": "FeatureCollection",
+        "features": items or [],
+    }
 
-        if fiona is None:
-            err_msg = "Cannot open AOI geospatial files without Fiona. " \
-                      "Please install the Fiona Python package."
-            self.eod.print_msg(err_msg, heading='warning')
-            self.logger.warning(err_msg)
-            return None
+    with open(output_file, "w", encoding="utf-8") as out_f:
+        json.dump(feature_collection, out_f, indent=2)
 
+    click.echo(f"Saved {len(feature_collection['features'])} item(s) to {output_file}")
+
+
+def make_aaa(username: Optional[str], password: Optional[str], environment: str):
+    if username and password:
+        return aaa.AAA_API(username, password, environment)
+    return None
+
+
+def make_dds(aaa_api, environment: str):
+    try:
+        return dds.DDS_API(aaa_api, environment=environment)
+    except TypeError:
+        return dds.DDS_API(aaa_api, environment)
+
+
+def make_search(aaa_api, environment: str):
+    try:
+        return search.Search_API(aaa_api=aaa_api, environment=environment)
+    except TypeError:
+        return search.Search_API(aaa_api, environment)
+
+
+def make_processes(aaa_api, environment: str):
+    try:
+        return Processes_API(aaa_api=aaa_api, environment=environment)
+    except TypeError:
+        return Processes_API(aaa_api, environment)
+
+
+def _normalize_status(value: Optional[str]) -> str:
+    if not value:
+        return ""
+    return "".join(ch for ch in str(value).upper() if ch.isalnum())
+
+
+def _extract_dds_status(item_info: Any) -> Optional[str]:
+    if not isinstance(item_info, dict):
+        return None
+
+    for key in ("status", "item_status", "itemStatus", "state"):
+        value = item_info.get(key)
+        if value is not None and str(value).strip():
+            return str(value)
+
+    props = item_info.get("properties")
+    if isinstance(props, dict):
+        for key in ("status", "item_status", "itemStatus", "state"):
+            value = props.get(key)
+            if value is not None and str(value).strip():
+                return str(value)
+
+    return None
+
+
+def _extract_dds_timestamp(item_info: Any) -> Optional[str]:
+    if not isinstance(item_info, dict):
+        return None
+
+    for key in (
+        "last_update",
+        "lastUpdate",
+        "timestamp",
+        "updated",
+        "updated_at",
+        "date",
+        "datetime",
+    ):
+        value = item_info.get(key)
+        if value is not None and str(value).strip():
+            return str(value).strip()
+
+    props = item_info.get("properties")
+    if isinstance(props, dict):
+        for key in (
+            "last_update",
+            "lastUpdate",
+            "timestamp",
+            "updated",
+            "updated_at",
+            "date",
+            "datetime",
+        ):
+            value = props.get(key)
+            if value is not None and str(value).strip():
+                return str(value).strip()
+
+    return None
+
+
+def _append_dds_retry_item(retry_file: str, collection: str, item_uuid: str, status: str,
+                           timestamp: Optional[str] = None) -> None:
+    retry_path = os.path.abspath(retry_file)
+    normalized_collection = str(collection or "").strip().lower()
+    normalized_uuid = str(item_uuid or "").strip().lower()
+    normalized_status = _normalize_status(status)
+    resolved_timestamp = timestamp or datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
+
+    rows: List[Tuple[str, Any]] = []
+    duplicate_found = False
+
+    # Refresh timestamp for existing retry rows with same collection/uuid/status.
+    if os.path.exists(retry_path):
         try:
-            aoi_data = parse_aoi_file(input_fn)
-            if aoi_data and len(aoi_data) > 0:
-                aoi_names = [d.get('name', 'unnamed') for d in aoi_data]
-                if len(aoi_data) > 1:
-                    info_msg = f"Loaded {len(aoi_data)} AOI(s): " \
-                              f"{', '.join(aoi_names)}"
-                    self.eod.print_msg(info_msg, heading='note')
-                    self.logger.info(info_msg)
-                    # Return the list of AOI dicts for multiple AOIs
-                    return aoi_data
-                else:
-                    # Return the WKT string for a single AOI
-                    return aoi_data[0]['wkt']
-            err_msg = f"Could not parse AOI file {input_fn}. " \
-                      f"No valid polygons found."
-            self.eod.print_msg(err_msg, heading='warning')
-            self.logger.warning(err_msg)
+            with open(retry_path, "r", encoding="utf-8") as in_f:
+                for line in in_f:
+                    candidate = line.strip()
+                    if not candidate:
+                        rows.append(("raw", ""))
+                        continue
+                    try:
+                        row = json.loads(candidate)
+                    except json.JSONDecodeError:
+                        rows.append(("raw", candidate))
+                        continue
+
+                    if not isinstance(row, dict):
+                        rows.append(("raw", candidate))
+                        continue
+
+                    row_collection = str(row.get("collection", "")).strip().lower()
+                    row_uuid = str(
+                        row.get("uuid")
+                        or row.get("id")
+                        or row.get("item_id")
+                        or row.get("itemId")
+                        or ""
+                    ).strip().lower()
+                    row_status = _normalize_status(str(row.get("status", "")))
+
+                    if (
+                        row_collection == normalized_collection
+                        and row_uuid == normalized_uuid
+                        and row_status == normalized_status
+                    ):
+                        row["timestamp"] = resolved_timestamp
+                        row["status"] = status
+                        duplicate_found = True
+
+                    rows.append(("json", row))
+
+            if duplicate_found:
+                retry_dir = os.path.dirname(retry_path)
+                if retry_dir:
+                    os.makedirs(retry_dir, exist_ok=True)
+
+                with open(retry_path, "w", encoding="utf-8") as out_f:
+                    for row_type, row_value in rows:
+                        if row_type == "json":
+                            out_f.write(json.dumps(row_value, ensure_ascii=True) + "\n")
+                        elif row_value == "":
+                            out_f.write("\n")
+                        else:
+                            out_f.write(str(row_value) + "\n")
+
+                return
+        except Exception:
+            # If dedupe scan/update fails for any reason, continue and append.
+            pass
+
+    record = {
+        "collection": collection,
+        "uuid": item_uuid,
+        "status": status,
+        "timestamp": resolved_timestamp,
+    }
+
+    retry_dir = os.path.dirname(retry_path)
+    if retry_dir:
+        os.makedirs(retry_dir, exist_ok=True)
+
+    with open(retry_path, "a", encoding="utf-8") as out_f:
+        out_f.write(json.dumps(record, ensure_ascii=True) + "\n")
+
+
+def download_dds_item(dds_api, collection: str, item_uuid: str, download_dir: str,
+                      queued_backoff_seconds: int = DEFAULT_DDS_BACKOFF_SECONDS,
+                      retry_file: str = DEFAULT_DDS_RETRY_FILE) -> Optional[Dict[str, Any]]:
+    waits = 0
+
+    while True:
+        item_info = dds_api.get_item(collection, item_uuid)
+        if item_info is None:
+            click.echo(f"Item not found: collection={collection}, uuid={item_uuid}")
             return None
-        except ValueError as e:
-            err_msg = f"Error parsing AOI file {input_fn}: {str(e)}"
-            self.eod.print_msg(err_msg, heading='warning')
-            self.logger.warning(err_msg)
-            return None
 
-    def ask_aoi(self, input_fn):
-        """
-        Asks the user for the geospatial input filename.
-        
-        :param input_fn: The geospatial input filename if already set by the
-                command-line.
-        :type  input_fn: str
-        
-        :return: The geospatial filename entered by the user or WKT string.
-        :rtype: str
-        """
+        status_value = _extract_dds_status(item_info)
+        normalized_status = _normalize_status(status_value)
 
-        if input_fn is None or input_fn == '':
-
-
-            if not self.eod.silent:
-                self.print_header("Enter Input Geospatial File or Feature")
-
-                msg = f"Enter the full path name of a " \
-                    f"{self.eod.var_colour}.gml{self.eod.reset_colour}, " \
-                    f"{self.eod.var_colour}.kml{self.eod.reset_colour}, " \
-                    f"{self.eod.var_colour}.shp{self.eod.reset_colour} or " \
-                    f"{self.eod.var_colour}.geojson{self.eod.reset_colour} " \
-                    f" containing an AOI or a WKT feature to " \
-                    f"restrict the search to a specific location"
-                err_msg = "No AOI or feature specified. Please enter a WKT " \
-                          "feature or a valid GML, KML, Shapefile or GeoJSON " \
-                          "file"
-                def_msg = "leave blank to exclude spatial filtering"
-                input_fn = self.get_input(msg, err_msg, required=False, 
-                                          def_msg=def_msg)
-
-        if input_fn is None or input_fn == '':
-            return None
-
-        if os.path.exists(input_fn):
-            input_fn = input_fn.strip()
-            input_fn = input_fn.strip("'")
-            input_fn = input_fn.strip('"')
-
-            # ---------------------------------
-            # Check validity of the input file
-            # ---------------------------------
-
-            input_fn = self.eod.validate_file(input_fn, True)
-
-            if not input_fn:
+        if normalized_status == "QUEUED":
+            if waits >= MAX_DDS_QUEUED_WAITS:
+                click.echo(
+                    f"Item remained Queued after {MAX_DDS_QUEUED_WAITS} wait(s): "
+                    f"collection={collection}, uuid={item_uuid}"
+                )
                 return None
 
-            # For AOI files, return the file path (not parsed data).
-            # Parsing happens later in search_order_download.
-            if any(input_fn.lower().endswith(ext)
-                   for ext in self.eod.aoi_extensions):
-                # Validate by parsing, but return the file path
-                self._parse_aoi_file(input_fn)
-                return input_fn
+            waits += 1
+            click.echo(
+                f"DDS item Queued: collection={collection}, uuid={item_uuid}. "
+                f"Waiting {queued_backoff_seconds}s before retry ({waits}/{MAX_DDS_QUEUED_WAITS})..."
+            )
+            time.sleep(queued_backoff_seconds)
+            continue
 
-        elif any(s in input_fn for s in self.eod.aoi_extensions):
-            err_msg = f"Input file {os.path.abspath(input_fn)} does not exist."
-            self.eod.print_msg(err_msg, heading="warning")
-            self.logger.warning(err_msg)
+        if normalized_status in {"ITEMRESTORING", "ITEMSRESTORING"}:
+            restoring_timestamp = _extract_dds_timestamp(item_info)
+            retry_abs_path = os.path.abspath(retry_file)
+            retry_rel_path = os.path.relpath(retry_abs_path, os.getcwd())
+            retry_rel_path = os.path.normpath(retry_rel_path)
+            if not (retry_rel_path.startswith(".") or os.path.isabs(retry_rel_path)):
+                retry_rel_path = f".{os.sep}{retry_rel_path}"
+            _append_dds_retry_item(
+                retry_file=retry_file,
+                collection=collection,
+                item_uuid=item_uuid,
+                status=status_value or "ItemRestoring",
+                timestamp=restoring_timestamp,
+            )
+            click.echo(
+                f"DDS item is {status_value or 'ItemRestoring'}; recorded for retry: "
+                f"collection={collection}, uuid={item_uuid}, file={retry_rel_path}"
+            )
+            click.echo(
+                "Retry with: "
+                f"py eodms_cli.py download --dds-retry-file \"{retry_rel_path}\""
+            )
+            return {
+                "_dds_retry_logged": True,
+                "collection": collection,
+                "uuid": item_uuid,
+                "status": status_value or "ItemRestoring",
+                "timestamp": restoring_timestamp,
+            }
+
+        if "download_url" not in item_info:
+            click.echo(
+                f"Item has no download URL: collection={collection}, uuid={item_uuid}, "
+                f"status={status_value or 'unknown'}"
+            )
             return None
 
-        else:
-            if not self.eod.eodms_geo.is_wkt(input_fn):
-                err_msg = "Input feature is not a valid WKT."
-                self.eod.print_msg(err_msg, heading="warning")
-                self.logger.warning(err_msg)
-                return None
+        dds_api.download_item(os.path.abspath(download_dir))
+        return item_info
 
-        return input_fn
 
-    def ask_aws(self, aws):
-        """
-        Asks the user if they'd like to download the image using AWS,
-            if applicable.
-        
-        :param aws: If already entered by the command-line, True if the user
-                    wishes to download from AWS.
-        :type  aws: boolean
-        
-        :return: True if the user wishes to download from AWS.
-        :rtype: boolean
-        """
+def _is_public_asset_collection(collection: str) -> bool:
+    normalized = str(collection or "").strip().lower()
+    return normalized in {"rcm-ard", "sentinel-1", "sentinel-2"}
 
-        if not aws:
 
-            if not self.eod.silent:
-                self.print_header("Download from AWS?")
+def _asset_filename(item_uuid: str, asset_name: str, href: str) -> str:
+    parsed = urlsplit(href)
+    file_name = os.path.basename(parsed.path)
+    file_name = unquote(file_name)
+    if file_name:
+        return file_name
+    return f"{item_uuid}_{asset_name}"
 
-                print("\nSome Radarsat-1 images contain direct download "
-                      "links to GeoTIFF files in an Open Data AWS "
-                      "Repository.")
 
-                msg = "For images that have an AWS link, would you like to " \
-                      "download the GeoTIFFs from the repository instead of " \
-                      "submitting an order to the EODMS?"
-                aws = click.confirm(msg, default=True)
+def _unique_destination_path(destination_dir: str, file_name: str) -> str:
+    full_path = os.path.join(destination_dir, file_name)
+    if not os.path.exists(full_path):
+        return full_path
 
-        return aws
+    stem, ext = os.path.splitext(file_name)
+    idx = 1
+    while True:
+        candidate = os.path.join(destination_dir, f"{stem}_{idx}{ext}")
+        if not os.path.exists(candidate):
+            return candidate
+        idx += 1
 
-    def ask_collection(self, coll, coll_lst=None):
-        """
-        Asks the user for the collection(s).
-        
-        :param coll: The collections if already set by the command-line.
-        :type  coll: str
-        :param coll_lst: A list of collections retrieved from the RAPI.
-        :type  coll_lst: list[str]
-        
-        :return: A list of collections entered by the user.
-        :rtype: list[str]
-        """
 
+def download_public_stac_assets(search_api, collection: str, item_uuid: str, download_dir: str) -> int:
+    item_info = search_api.get_item(collection, item_uuid)
+    if item_info is None:
+        click.echo(f"Item not found: collection={collection}, uuid={item_uuid}")
+        return 0
+
+    assets = item_info.get("assets") if isinstance(item_info, dict) else None
+    if not isinstance(assets, dict) or not assets:
+        click.echo(f"Item has no assets: collection={collection}, uuid={item_uuid}")
+        return 0
+
+    destination = os.path.abspath(download_dir)
+    os.makedirs(destination, exist_ok=True)
+
+    downloaded_count = 0
+    for asset_name, asset_obj in assets.items():
+        href = None
+        if isinstance(asset_obj, dict):
+            href = asset_obj.get("href")
+        elif isinstance(asset_obj, str):
+            href = asset_obj
+
+        if not href:
+            continue
+
+        out_name = _asset_filename(item_uuid, str(asset_name), str(href))
+        out_path = _unique_destination_path(destination, out_name)
+
+        click.echo(f"Downloading asset '{asset_name}'...")
+        try:
+            with urlopen(str(href)) as resp, open(out_path, "wb") as out_f:
+                while True:
+                    chunk = resp.read(1024 * 1024)
+                    if not chunk:
+                        break
+                    out_f.write(chunk)
+        except Exception as exc:
+            click.echo(f"Failed to download asset '{asset_name}': {exc}")
+            continue
+
+        downloaded_count += 1
+        click.echo(f"Saved asset to {out_path}")
+
+    if downloaded_count == 0:
+        click.echo(f"No downloadable asset links found for collection={collection}, uuid={item_uuid}")
+
+    return downloaded_count
+
+
+def _load_download_items_from_geojson(input_file: str) -> List[Dict[str, Any]]:
+    try:
+        with open(input_file, "r", encoding="utf-8") as src:
+            raw_text = src.read().strip()
+    except Exception as exc:
+        raise click.ClickException(f"Failed to read input file '{input_file}': {exc}")
+
+    if not raw_text:
+        return []
+
+    payload: Any = None
+    parsed_json = False
+    try:
+        payload = json.loads(raw_text)
+        parsed_json = True
+    except json.JSONDecodeError:
+        parsed_json = False
+
+    if parsed_json:
+        if isinstance(payload, dict):
+            features = payload.get("features")
+            if isinstance(features, list):
+                return [feature for feature in features if isinstance(feature, dict)]
+
+            retry_items = payload.get("items")
+            if isinstance(retry_items, list):
+                return [item for item in retry_items if isinstance(item, dict)]
+
+            # Accept a single retry object (for example one JSONL line copied to .json)
+            # when it has an identifiable item UUID/id field.
+            if _extract_item_uuid(payload):
+                return [payload]
+
+            raise click.ClickException(
+                "Input JSON must be GeoJSON FeatureCollection ('features') "
+                "or retry JSON with an 'items' list (or a single item object with uuid/id)."
+            )
+
+        if isinstance(payload, list):
+            return [item for item in payload if isinstance(item, dict)]
+
+        raise click.ClickException("Input JSON must be an object or list.")
+
+    # Fallback: JSON lines support for retry files (one object per line).
+    items: List[Dict[str, Any]] = []
+    for idx, line in enumerate(raw_text.splitlines(), start=1):
+        candidate = line.strip()
+        if not candidate:
+            continue
+        try:
+            parsed_line = json.loads(candidate)
+        except json.JSONDecodeError as exc:
+            raise click.ClickException(
+                f"Invalid JSON line at {idx} in '{input_file}': {exc}"
+            )
+        if isinstance(parsed_line, dict):
+            items.append(parsed_line)
+
+    if items:
+        return items
+
+    raise click.ClickException(
+        "Input file must be GeoJSON, JSON retry records, or JSONL retry records."
+    )
+
+
+def parse_legacy_order_items(order_items: str) -> Tuple[List[str], List[str]]:
+    """Parse legacy order-item selector syntax: order:id1,id2|item:id3,id4"""
+    order_ids: List[str] = []
+    item_ids: List[str] = []
+
+    for part in order_items.split("|"):
+        part = part.strip()
+        if not part or ":" not in part:
+            continue
+
+        key, raw_ids = [p.strip().lower() for p in part.split(":", 1)]
+        ids = [value.strip() for value in raw_ids.split(",") if value.strip()]
+
+        if key.startswith("order"):
+            order_ids.extend(ids)
+        elif key.startswith("item"):
+            item_ids.extend(ids)
+
+    return order_ids, item_ids
+
+
+def _to_rapi_orders_dt_string(value: Any) -> Optional[str]:
+    if value is None:
+        return None
+
+    if isinstance(value, datetime):
+        dt_val = value.astimezone(timezone.utc)
+        return dt_val.strftime("%Y-%m-%dT%H:%M:%SZ")
+
+    if isinstance(value, str):
+        parsed = _parse_iso_datetime(value)
+        if parsed is not None:
+            return parsed.strftime("%Y-%m-%dT%H:%M:%SZ")
+        stripped = value.strip()
+        if stripped:
+            return stripped
+
+    return None
+
+
+def _build_rapi_orders_url(rapi_api: EODMSRAPI, max_orders: int,
+                           dtstart: Any = None, dtend: Any = None,
+                           status: Optional[str] = None,
+                           out_format: str = "json") -> str:
+    params: Dict[str, Any] = {"maxOrders": max_orders}
+
+    dtstart_s = _to_rapi_orders_dt_string(dtstart)
+    dtend_s = _to_rapi_orders_dt_string(dtend)
+    if dtstart_s:
+        params["dtstart"] = dtstart_s
+    if dtend_s:
+        params["dtend"] = dtend_s
+    if status:
+        params["status"] = status.upper()
+
+    param_str = urlencode(params)
+    return f"{rapi_api.rapi_root}/order?{param_str}&format={out_format}"
+
+
+def _safe_rapi_call(callable_obj, *args, **kwargs):
+    result = callable_obj(*args, **kwargs)
+    if isinstance(result, QueryError):
+        raise click.ClickException(result.get_msgs(as_str=True))
+    return result
+
+
+def _first_dict(payload: Any) -> Optional[Dict[str, Any]]:
+    if isinstance(payload, dict):
+        if isinstance(payload.get("results"), list):
+            for entry in payload["results"]:
+                found = _first_dict(entry)
+                if found is not None:
+                    return found
+        if isinstance(payload.get("items"), list):
+            for entry in payload["items"]:
+                found = _first_dict(entry)
+                if found is not None:
+                    return found
+        return payload
+    if isinstance(payload, list):
+        for entry in payload:
+            found = _first_dict(entry)
+            if found is not None:
+                return found
+    return None
+
+
+def _extract_first_order_id(payload: Any) -> Optional[str]:
+    if isinstance(payload, dict):
+        for key in ("orderId", "order_id", "ORDER_ID"):
+            value = payload.get(key)
+            if value is not None:
+                return str(value)
+        for key in ("results", "items"):
+            value = payload.get(key)
+            if isinstance(value, list):
+                for entry in value:
+                    found = _extract_first_order_id(entry)
+                    if found:
+                        return found
+        return None
+
+    if isinstance(payload, list):
+        for entry in payload:
+            found = _extract_first_order_id(entry)
+            if found:
+                return found
+
+    return None
+
+
+def _extract_order_key(stac_item: Dict[str, Any]) -> Optional[str]:
+    props = stac_item.get("properties") if isinstance(stac_item, dict) else None
+    if not isinstance(props, dict):
+        return None
+
+    return (
+        props.get("order_key")
+        or props.get("orderKey")
+        or props.get("Order Key")
+    )
+
+
+def _extract_item_uuid(stac_item: Dict[str, Any]) -> Optional[str]:
+    if not isinstance(stac_item, dict):
+        return None
+
+    for key in ("id", "uuid", "item_id", "itemId"):
+        value = stac_item.get(key)
+        if value is not None and str(value).strip():
+            return str(value)
+
+    props = stac_item.get("properties")
+    if isinstance(props, dict):
+        for key in ("uuid", "UUID", "id"):
+            value = props.get(key)
+            if value is not None and str(value).strip():
+                return str(value)
+
+    return None
+
+
+def _parse_iso_datetime(value: Optional[str]) -> Optional[datetime]:
+    if not value or not isinstance(value, str):
+        return None
+
+    raw = value.strip()
+    if not raw:
+        return None
+
+    # Handle STAC UTC suffix format.
+    if raw.endswith("Z"):
+        raw = raw[:-1] + "+00:00"
+
+    try:
+        dt_val = datetime.fromisoformat(raw)
+    except ValueError:
+        return None
+
+    if dt_val.tzinfo is None:
+        dt_val = dt_val.replace(tzinfo=timezone.utc)
+
+    return dt_val.astimezone(timezone.utc)
+
+
+def _format_rapi_datetime(dt_val: datetime) -> str:
+    return dt_val.strftime("%Y%m%d_%H%M%S")
+
+
+def _build_rapi_dates_from_stac_item(stac_item: Dict[str, Any], span_days: int = 1) -> Optional[List[Dict[str, str]]]:
+    props = stac_item.get("properties") if isinstance(stac_item, dict) else None
+    if not isinstance(props, dict):
+        return None
+
+    start_dt = _parse_iso_datetime(
+        props.get("start_datetime")
+        or props.get("startDateTime")
+        or props.get("acquisition_start")
+        or props.get("acquisitionStart")
+    )
+    end_dt = _parse_iso_datetime(
+        props.get("end_datetime")
+        or props.get("endDateTime")
+        or props.get("acquisition_end")
+        or props.get("acquisitionEnd")
+    )
+
+    if start_dt is None and end_dt is None:
+        center_dt = _parse_iso_datetime(
+            props.get("datetime")
+            or props.get("acquisition_datetime")
+            or props.get("acquisitionDateTime")
+        )
+        if center_dt is None:
+            return None
+        start_dt = center_dt
+        end_dt = center_dt + timedelta(days=span_days)
+    elif start_dt is None and end_dt is not None:
+        start_dt = end_dt - timedelta(days=span_days)
+    elif end_dt is None and start_dt is not None:
+        end_dt = start_dt + timedelta(days=span_days)
+
+    if start_dt is None or end_dt is None:
+        return None
+
+    if end_dt < start_dt:
+        start_dt, end_dt = end_dt, start_dt
+
+    return [{"start": _format_rapi_datetime(start_dt), "end": _format_rapi_datetime(end_dt)}]
+
+
+def _download_rapi_items(rapi_api: EODMSRAPI, items: List[Dict[str, Any]], download_dir: str) -> None:
+    if not items:
+        click.echo("No downloadable items found.")
+        return
+
+    destination = os.path.abspath(download_dir)
+    os.makedirs(destination, exist_ok=True)
+
+    click.echo(f"Downloading {len(items)} item(s) to {destination}")
+    result = _safe_rapi_call(rapi_api.download, items, destination)
+
+    if isinstance(result, list):
+        click.echo(f"Downloaded/attempted {len(result)} item(s).")
+    else:
+        click.echo("Download request submitted.")
+
+
+def _load_json_input(raw: Optional[str], label: str) -> Optional[Dict[str, Any]]:
+    if raw is None:
+        return None
+
+    candidate = raw.strip()
+    if not candidate:
+        return None
+
+    if os.path.exists(candidate):
+        with open(candidate, "r", encoding="utf-8") as src:
+            return json.load(src)
+
+    try:
+        return json.loads(candidate)
+    except json.JSONDecodeError as exc:
+        raise click.ClickException(
+            f"Invalid {label} JSON. Provide either a valid JSON string or a file path."
+        ) from exc
+
+
+def _print_process_summary(processes_json: Dict[str, Any]) -> None:
+    all_processes = processes_json.get("processes", [])
+    click.echo("\nProcessing Service:\n")
+    for process_obj in all_processes:
+        process_id = process_obj.get("id", "N/A")
+        version = process_obj.get("version", "N/A")
+        description = process_obj.get("description") or process_obj.get("abstract") or "N/A"
+        click.echo(f"* {process_id} (v{version}): {description}")
+
+    click.echo("\nSAR Toolbox:\n")
+    click.echo(
+        "* SAR_Toolbox (vX.X): Filters, Ortho-rectification and mosaic "
+        "Radiometry, Polarimetry, Interferometry, Analysis Ready Data. Support for RADARSAT-2, RCMImageProducts"
+    )
+
+
+def _example_scalar_from_schema(schema: Dict[str, Any], input_name: str) -> Any:
+    if not isinstance(schema, dict):
+        return "example"
+
+    if "default" in schema:
+        return schema["default"]
+
+    enum_vals = schema.get("enum")
+    if isinstance(enum_vals, list) and enum_vals:
+        return enum_vals[0]
+
+    schema_type = schema.get("type")
+    schema_format = schema.get("format")
+
+    if schema_type == "boolean":
+        return True
+    if schema_type in ("integer", "number"):
+        return 1
+    if schema_type == "array":
+        item_schema = schema.get("items")
+        if isinstance(item_schema, dict):
+            return [_example_scalar_from_schema(item_schema, input_name)]
+        return ["example"]
+    if schema_type == "object":
+        return {}
+
+    if schema_format == "date-time":
+        return "2000-01-01T00:00:00Z"
+
+    lower_name = input_name.lower()
+    if lower_name in ("uuid", "id", "segment_id"):
+        return "00000000-0000-0000-0000-000000000000"
+    if lower_name in ("start_time", "stop_time"):
+        return "2000-01-01T00:00:00Z"
+
+    return "example"
+
+
+def _example_value_from_input_def(input_name: str, input_def: Dict[str, Any]) -> Any:
+    if not isinstance(input_def, dict):
+        return "example"
+
+    if "default" in input_def:
+        return input_def["default"]
+
+    for schema_key in ("schema", "valueSchema"):
+        schema_obj = input_def.get(schema_key)
+        if isinstance(schema_obj, dict):
+            return _example_scalar_from_schema(schema_obj, input_name)
+
+    return _example_scalar_from_schema(input_def, input_name)
+
+
+def _build_sample_payload(process_id: str, process_json: Dict[str, Any]) -> Dict[str, Any]:
+    input_defs = process_json.get("inputs", {})
+    sample_inputs: Dict[str, Any] = {}
+
+    if isinstance(input_defs, dict):
+        for input_name, input_def in input_defs.items():
+            sample_inputs[input_name] = _example_value_from_input_def(str(input_name), input_def)
+
+    sample_outputs = {
+        f"{process_id}-response": {
+            "format": {"mediaType": "application/json"}
+        }
+    }
+
+    return {
+        "inputs": sample_inputs,
+        "outputs": sample_outputs,
+        "mode": "async",
+    }
+
+
+@click.group(cls=OrderedHelpGroup, context_settings={"help_option_names": ["-h", "--help"]})
+def cli():
+    """EODMS CLI v2: STAC/DDS-first, non-interactive with minimal legacy ports."""
+    click.echo()
+
+
+@cli.command("configure")
+@click.option("--username", "username", "-u", required=True,
+              help="EODMS username to store in config.ini.")
+@click.option("--password", "password", "-p", required=True,
+              help="EODMS password to store in config.ini.")
+def configure_cmd(username: str, password: str):
+    """Save Credentials.username/password to %USERPROFILE%\\.eodms\\config.ini."""
+
+    cfg_path = _save_credentials_to_config(username=username, password=password)
+    click.echo(f"Saved credentials to {cfg_path}")
+
+
+@cli.command("search")
+@click.option("--username", "-u", required=False, help="EODMS username.")
+@click.option("--password", "-p", required=False, help="EODMS password.")
+@click.option("--collection", "-c", required=False, help="Collection name.")
+@click.option("--list", "list_collections", is_flag=True,
+              help="List available STAC collections and exit.")
+@click.option("--uuid2record", "uuid2record", is_flag=True,
+              help="Resolve UUID to order_key (search) then recordId (RAPI).")
+@click.option("--uuid", required=False, default=None,
+              help="UUID (or comma-separated UUIDs) used with --uuid2record.")
+@click.option("--queryables", "show_queryables", is_flag=True,
+              help="Print queryables for --collection and exit.")
+@click.option("--datetime", "-d", "datetime_range", required=False, default=None,
+              help='Temporal filter as ISO 8601 string/range (example: "2023-01-01/2023-12-31").')
+@click.option("--bbox", "-b", required=False, default=None,
+              help="Bounding box as west,south,east,north")
+@click.option("--limit", "-l", required=False, default=1000, type=int,
+              help="Maximum number of items to fetch (default: 1000).")
+@click.option("--filter", "-f", "filter_text", required=False, default=None,
+              help="CQL2 text filter expression.")
+@click.option("--s-intersect", "s_intersect", required=False, default=None,
+              help="WKT geometry used with S_INTERSECTS.")
+@click.option("--aoi", required=False, default=None, type=click.Path(exists=True),
+              help="Path to geospatial AOI file with 1-5 polygons.")
+@click.option("--output", "-o", required=False, default=None,
+              help="Output GeoJSON file for search results.")
+@click.option("--env", "-e", required=False, default="prod",
+              help='Environment (default: "prod").')
+@handle_service_errors
+def search_cmd(
+    username: Optional[str],
+    password: Optional[str],
+    collection: Optional[str],
+    list_collections: bool,
+    uuid2record: bool,
+    uuid: Optional[str],
+    show_queryables: bool,
+    datetime_range: Optional[str],
+    bbox: Optional[str],
+    limit: int,
+    filter_text: Optional[str],
+    s_intersect: Optional[str],
+    aoi: Optional[str],
+    output: Optional[str],
+    env: str,
+):
+    """STAC geotemporal/queryables and GeoJSON output."""
+
+    username, password = resolve_credentials(username, password)
+
+    bbox_list = parse_bbox(bbox)
+
+    aaa_api = make_aaa(username, password, env)
+
+    if list_collections:
+        search_api = make_search(aaa_api, env)
+        collections = []
+        for coll in search_api.client.get_collections():
+            coll_title = getattr(coll, "title", None) or coll.id
+            collections.append((coll.id, coll_title))
+
+        if not collections:
+            click.echo("No collections found.")
+            return
+
+        click.echo(f"Found {len(collections)} collection(s):")
+        for coll_id, coll_title in collections:
+            if coll_title == coll_id:
+                click.echo(f"- {coll_id}")
+            else:
+                click.echo(f"- {coll_id}: {coll_title}")
+        return
+
+    if uuid2record:
+        if not uuid:
+            raise click.ClickException("--uuid is required with --uuid2record.")
+        if not collection:
+            raise click.ClickException("--collection is required with --uuid2record.")
+        if not username or not password:
+            raise click.ClickException("--username and --password are required with --uuid2record.")
+
+        search_api = make_search(aaa_api, env)
+        rapi_api = EODMSRAPI(username, password)
+        uuid_values = [u.strip() for u in uuid.split(",") if u.strip()]
+        if not uuid_values:
+            raise click.ClickException("At least one UUID must be provided.")
+
+        for uuid_value in uuid_values:
+            item = search_api.get_item(collection, uuid_value)
+            if item is None:
+                click.echo(f"{uuid_value}: no item found in search/{collection}")
+                continue
+
+            order_key = _extract_order_key(item)
+            if not order_key:
+                click.echo(f"{uuid_value}: no order_key on STAC item")
+                continue
+
+            filter_dict = {"ARCHIVE_IMAGE.ORDER_KEY": ("=", [order_key])}
+            rapi_dates = _build_rapi_dates_from_stac_item(item, span_days=1)
+            try:
+                search_kwargs: Dict[str, Any] = {
+                    "filters": filter_dict,
+                    "max_results": 5,
+                }
+                if rapi_dates:
+                    search_kwargs["dates"] = rapi_dates
+
+                _safe_rapi_call(rapi_api.search, collection, **search_kwargs)
+                payload = _safe_rapi_call(rapi_api.get_results, "brief", show_progress=False)
+            except Exception as exc:
+                click.echo(f"{uuid_value}: order_key={order_key}; RAPI lookup error: {exc}")
+                continue
+
+            record = _first_dict(payload)
+            record_id = None
+            if isinstance(record, dict):
+                record_id = (
+                    record.get("recordId")
+                    or record.get("record_id")
+                    or record.get("RECORD_ID")
+                )
+
+            if record_id:
+                click.echo(f"{uuid_value}: order_key={order_key}; record_id={record_id}")
+            else:
+                click.echo(f"{uuid_value}: order_key={order_key}; record_id not found")
+        return
+
+    if show_queryables:
+        if not collection:
+            raise click.ClickException("--collection is required with --queryables.")
+
+        search_api = make_search(aaa_api, env)
+        coll = search_api.client.get_collection(collection)
         if coll is None:
+            raise click.ClickException(f"Collection not found: {collection}")
 
-            if coll_lst is None:
-                coll_lst = self.eod.get_collections(True)
+        search_api.print_queryables(coll)
+        return
 
-            if self.eod.silent:
-                err_msg = "No collection specified. Exiting process."
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
+    if not collection:
+        raise click.ClickException("--collection is required unless --list is used.")
 
-            self.print_header("Enter Collection")
-
-            # List available collections for this user
-            print("\nAvailable Collections:\n")
-            coll_lst = sorted(coll_lst, key=lambda x: x['title'])
-            for idx, c in enumerate(coll_lst):
-                msg = f"{self.eod.var_colour}{idx + 1}{self.eod.reset_colour}" \
-                    f". {c['title']} ({c['id']})"
-                print(self.wrap_text(msg))
-
-            # Prompted user for number(s) from list
-            msg = "Enter the number of a collection from the list " \
-                  "above (for multiple collections, enter each number " \
-                  "separated with a comma)"
-            err_msg = "At least one collection must be specified."
-            in_coll = self.get_input(msg, err_msg)
-
-            # Convert number(s) to collection name(s)
-            coll_vals = in_coll.split(',')
-
-            # ---------------------------------------
-            # Check validity of the collection entry
-            # ---------------------------------------
-
-            check = self.eod.validate_int(coll_vals, len(coll_lst))
-            if not check:
-                err_msg = "A valid Collection must be specified. " \
-                          "Exiting process."
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-
-            coll = [coll_lst[int(i) - 1]['id'] for i in coll_vals
-                    if i.isdigit()]
-        else:
-            coll = coll.split(',')
-
-        # ------------------------------
-        # Check validity of Collections
-        # ------------------------------
-        for c in coll:
-            check = self.eod.validate_collection(c)
-            if not check:
-                err_msg = f"Collection '{c}'' is not valid."
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-
-        return coll
-
-    def ask_dates(self, dates):
-        """
-        Asks the user for dates.
-        
-        :param dates: The dates if already set by the command-line.
-        :type  dates: str
-        
-        :return: The dates entered by the user.
-        :rtype: str
-        """
-
-        # Get the date range
-        if dates is None:
-
-            if not self.eod.silent:
-                self.print_header("Enter Date Range")
-
-                msg = f"Enter a date range (ex: " \
-                    f"{self.eod.var_colour}20200525-20200630T200950" \
-                    f"{self.eod.reset_colour}) or a previous time-frame " \
-                    f"({self.eod.var_colour}24 hours{self.eod.reset_colour})"
-                def_msg = "leave blank to search all years"
-                dates = self.get_input(msg, required=False, def_msg=def_msg)
-
-        # -------------------------------
-        # Check validity of filter input
-        # -------------------------------
-        if dates is not None and not dates == '':
-            dates = self.eod.validate_dates(dates)
-
-            if not dates:
-                err_msg = "The dates entered are invalid. "
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-
-        return dates
-
-    def ask_fields(self, csv_fields, fields):
-
-        if csv_fields is not None:
-            return csv_fields.split(',')
-
-        srch_fields = []
-        for f in fields:
-            if f.lower() in self.eod.csv_unique:
-                srch_fields.append(f.lower())
-
-        if len(srch_fields) > 0:
-            return srch_fields
-
-        if not self.eod.silent:
-            self.print_header("Enter CSV Unique Fields")
-
-            print("\nAvailable fields in the CSV file:")
-            for f in fields:
-                print(f"  {f}")
-
-            msg = "Enter the fields from the CSV file which can be used to " \
-                  "determine the images (separate each with a comma)"
-            input_fields = self.get_input(msg)
-
-            srch_fields = [f.strip() for f in input_fields.split(',')]
-
-            return srch_fields
-
-    def ask_filter(self, filters):
-        """
-        Asks the user for the search filters.
-        
-        :param filters: The filters if already set by the command-line.
-        :type  filters: str
-        
-        :return: A dictionary containing the filters entered by the user.
-        :rtype: dict
-        """
-
-        if filters is None:
-            filt_dict = {}
-
-            if not self.eod.silent:
-
-                self.print_header("Enter Filters")
-
-                # Ask for the filters for the given collection(s)
-                for coll in self.params['collections']:
-                    coll_id = self.eod.get_full_collid(coll)
-
-                    av_fields = self.eod.get_available_fields(coll_id,
-                                                               'title')
-                    if av_fields is None:
-                        err_msg = f"Could not retrieve available fields " \
-                                  f"for '{coll_id}'."
-                        self.eod.print_msg(err_msg, heading='error')
-                        self.logger.error(err_msg)
-                        self.eod.exit_cli(1)
-
-                    print(f"\nAvailable fields for '{coll}' (STAC):")
-                    printed_queryables = self.eod.print_queryables(coll_id)
-                    if not printed_queryables:
-                        avail_fields = sorted(av_fields.get('results', {}).keys())
-                        fields_str = ', '.join(avail_fields)
-                        print(self.wrap_text(fields_str, init_indent='  '))
-
-                    msg = "Enter the CQL2 filter to apply to the search e.g. beam_mnemonic LIKE 'SC30M%' AND relative_orbit = 10"
-                    def_msg = "leave blank for no filter"
-                    filt_items = self.get_input(msg, required=False,
-                                                def_msg=def_msg)
-
-                    filt_dict[coll_id] = filt_items if filt_items else ''
-
-        else:
-            # User specified in command-line
-
-            # Possible formats:
-            #   1. Only one collection: <field_id>=<value>|<value>,
-            #       <field_id>=<value>&<value>,...
-            #   2. Multiple collections but only specifying one set of filters:
-            #       <coll_id>.<field_id>=<value>|<value>,...
-            #   3. Multiple collections with filters:
-            #       <coll_id>.<field_id>=<value>,...
-            #       <coll_id>.<field_id>=<value>,...
-
-            filt_dict = {}
-
-            for coll in self.params['collections']:
-                # Split filters by comma
-                filt_lst = filters.split(',')
-                for f in filt_lst:
-                    f = f.strip('"')
-                    if f == '':
-                        continue
-                    if f.find('.') > -1:
-                        coll, filt_items = f.split('.')
-                        filt_items = self.eod.validate_filters(filt_items,
-                                                               coll)
-                        if not filt_items:
-                            self.eod.exit_cli(1)
-                        coll_id = self.eod.get_full_collid(coll)
-                        if coll_id in filt_dict.keys():
-                            coll_filters = filt_dict.get(coll_id)
-                        else:
-                            coll_filters = []
-                        coll_filters.append(
-                            filt_items.replace('"', '').replace("'", ''))
-                        filt_dict[coll_id] = coll_filters
-                    else:
-                        coll_id = self.eod.get_collid_by_name(coll)
-                        if coll_id in filt_dict.keys():
-                            coll_filters = filt_dict[coll_id]
-                        else:
-                            coll_filters = []
-                        coll_filters.append(f)
-                        filt_dict[coll_id] = coll_filters
-
-        return filt_dict
-
-    def ask_input_file(self, input_fn, msg, file_type='csv'):
-        """
-        Asks the user for the input filename.
-        
-        :param input_fn: The input filename if already set by the command-line.
-        :type  input_fn: str
-        :param msg: The message used to ask the user.
-        :type  msg: str
-        
-        :return: The input filename.
-        :rtype: str
-        """
-
-        if input_fn is None or input_fn == '':
-
-            if self.eod.silent:
-                err_msg = f"No {file_type.upper()} file specified. " \
-                            f"Exiting process."
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-
-            self.print_header(f"Enter Input {file_type.upper()} File")
-
-            err_msg = f"No {file_type.upper()} specified. " \
-                        f"Please enter a valid {file_type.upper()} file"
-            input_fn = self.get_input(msg, err_msg)
-
-        if not os.path.exists(input_fn):
-            err_msg = f"The specified {file_type.upper()} file ({input_fn}) " \
-                        f"does not exist. Please enter a valid " \
-                        f"{file_type.upper()} file."
-            self.eod.print_msg(err_msg, heading='error')
-            self.logger.error(err_msg)
-            self.eod.exit_cli(1)
-
-        return input_fn
-
-    def ask_maximum(self, maximum, max_type='order'):
-        """
-        Asks the user for maximum number of order items and the number of
-            items per order.
-        
-        :param maximum: The maximum if already set by the command-line.
-        :type  maximum: str
-        :param max_type: The type of maximum to set ('order' or 'download').
-        :type  max_type: str
-        :param no_order: Determines whether the maximum is for searching or
-        ordering.
-        :type  no_order: boolean
-        
-        :return: If max_type is 'order', the maximum number of order items
-        and/or number of items per order, separated by ':'. If max_type is
-        'download', a single number specifying how many images to download.
-        :rtype: str
-        """
-
-        # Get the no_order value
-        no_order = self.params.get('no_order')
-
-        if maximum is None or maximum == '':
-
-            if not self.eod.silent:
-                if no_order:
-                    self.print_header("Enter Maximum Search Results")
-                    msg = "Enter the maximum number of images you would " \
-                          "like to search for"
-                    def_msg = "leave blank to search for all images"
-
-                    maximum = self.get_input(msg, required=False, 
-                                             def_msg=def_msg)
-
-                    return maximum
-
-                if max_type == 'download':
-                    self.print_header("Enter Maximum for Downloads")
-                    msg = "Enter the number of images with status " \
-                          "AVAILABLE_FOR_DOWNLOAD you would like to " \
-                          "download"
-                    def_msg = "leave blank to download all images with " \
-                        "this status"
-
-                    maximum = self.get_input(msg, required=False, 
-                                             def_msg=def_msg)
-
-                    return maximum
-                else:
-                    if not self.process == 'order_csv':
-
-                        self.print_header("Enter Maximums for Ordering")
-
-                        msg = "Enter the total number of images you'd " \
-                              "like to order"
-                        def_msg = "leave blank for no limit"
-                        total_records = self.get_input(msg, required=False, 
-                                                       def_msg=def_msg)
-
-                        # ------------------------------------------
-                        # Check validity of the total_records entry
-                        # ------------------------------------------
-
-                        if total_records == '':
-                            total_records = None
-                        else:
-                            total_records = self.eod.validate_int(total_records)
-                            if not total_records:
-                                self.eod.print_msg("Total number of images "
-                                                   "value not valid. "
-                                                   "Excluding it.",
-                                                   indent=False, 
-                                                   heading='warning')
-                                total_records = None
-                            else:
-                                total_records = str(total_records)
-                    else:
-                        total_records = None
-
-                    maximum = ':'.join(filter(None, [total_records]))
-
-        else:
-
-            if max_type == 'order':
-                if self.process == 'order_csv':
-
-                    self.print_header("Enter Images per Order")
-
-                    if maximum.find(':') > -1:
-                        total_records, order_limit = maximum.split(':')
-                    else:
-                        total_records = None
-                        order_limit = maximum
-
-                    maximum = ':'.join(filter(None, [total_records,
-                                                     order_limit]))
-
-        return maximum
-
-    def ask_orderitems(self, orderitems):
-        """
-        Asks the user for a list Order IDs or Order Item IDs.
-
-        :param orderitems
-
-        """
-
-        if orderitems is None:
-            if not self.eod.silent:
-                self.print_header("Order/Order Item IDs")
-
-                msg = "\nEnter a list of Order IDs and/or Order Item IDs, " \
-                      "separating each ID with a comma and separating Order " \
-                      "IDs and Order Items with a vertical line " \
-                      "(ex: 'orders:<order_id>,<order_id>|items:" \
-                      "<order_item_id>,...')"
-                def_msg = "leave blank to skip"
-                orderitems = self.get_input(msg, required=False, 
-                                            def_msg=def_msg)
-
-        return orderitems
-
-    def ask_order(self, no_order):
-        """
-        Asks the user if they would like to suppress ordering and downloading.
-
-        :param no_order:
-        :return:
-        """
-
-        if no_order is None:
-            if not self.eod.silent:
-                self.print_header("Suppress Ordering")
-
-                msg = "Would you like to only search and not order?"
-                no_order = click.confirm(msg, default=False)
-
-        return no_order
-
-    def ask_output(self, output):
-        """
-        Asks the user for the output file.
-        
-        :param output: The output if already set by the command-line.
-        :type  output: str
-        
-        :return: The output filename.
-        :rtype: str
-        """
-
-        if output is None:
-
-            if not self.eod.silent:
-                self.print_header("Enter Output File")
-
-                msg = f"\nEnter the full path of the output file. " \
-                      f"For geospatial, use extension " \
-                      f"{self.eod.var_colour}.geojson{self.eod.reset_colour}," \
-                      f" {self.eod.var_colour}.kml{self.eod.reset_colour}, " \
-                      f"{self.eod.var_colour}.gml{self.eod.reset_colour}, or" \
-                      f" {self.eod.var_colour}.shp{self.eod.reset_colour}; " \
-                      f"for a list, use {self.eod.var_colour}." \
-                      f"csv{self.eod.reset_colour}"
-                if self.params.get('no_order', False):
-                    msg += " (NOTE: ordering is suppressed so no results information will be available without an output file)"
-                def_msg = "default is no output file"
-                output = self.get_input(msg, required=False, 
-                                        def_msg=def_msg)
-
-        return output
-
-    def ask_overlap(self, overlap):
-
-        if overlap is None:
-
-            if not self.eod.silent:
-                self.print_header("Enter Minimum Overlap Percentage")
-
-                msg = "\nEnter the minimum percentage of overlap between " \
-                      "images and the AOI"
-                def_msg = "leave blank for no overlap limit"
-                overlap = self.get_input(msg, required=False, def_msg=def_msg)
-
-        return overlap
-
-    def ask_priority(self, priority):
-        """
-        Asks the user for the order priority level
-        
-        :param priority: The priority if already set by the command-line.
-        :type  priority: str
-        
-        :return: The priority level.
-        :rtype: str
-        """
-
-        priorities = ['low', 'medium', 'high', 'urgent']
-
-        if priority is None:
-            if not self.eod.silent:
-                self.print_header("Enter Priority")
-
-                msg = "Enter the priority level for the order"
-
-                priority = self.get_input(msg, required=False,
-                                          options=priorities, default='medium')
-
-        if priority is None or priority == '':
-            priority = 'Medium'
-        elif priority.lower() not in priorities:
-            self.eod.print_msg("Not a valid 'priority' entry. "
-                               "Setting priority to 'Medium'.", indent=False, 
-                               heading='warning')
-            priority = 'Medium'
-
-        return priority
-
-    def ask_process(self):
-        """
-        Asks the user what process they would like to run.
-        
-        :return: The value the process the user has chosen.
-        :rtype: str
-        """
-
-        if self.eod.silent:
-            process = 'full'
-        else:
-            self.print_header("Choose Process Option")
-            choice_strs = []
-            for idx, v in enumerate(proc_choices.items()):
-                desc_str = re.sub(r'\s+', ' ', v[1]['desc'].replace('\n', ''))
-                choice_strs.append(self.wrap_text(f"{self.eod.var_colour}{idx + 1}" \
-                                    f"{self.eod.reset_colour}: ({v[0]}) " \
-                                    f"{desc_str}", sub_indent='     '))
-            choices = '\n'.join(choice_strs)
-
-            print(f"\nWhat would you like to do?\n\n{choices}")
-            msg = "Please choose the type of process"
-            process = self.get_input(msg, required=False, default='1')
-
-            if self.testing:
-                print(f"FOR TESTING - Process entered: {process}")
-
-            if process == '':
-                process = 'full'
-            else:
-                # Set process value and check its validity
-
-                process = self.eod.validate_int(process)
-
-                if not process:
-                    err_msg = "Invalid value entered for the 'process' " \
-                              "parameter."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-
-                if process > len(proc_choices.keys()):
-                    err_msg = "Invalid value entered for the 'process' " \
-                              "parameter."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-                else:
-                    process = list(proc_choices.keys())[int(process) - 1]
-
-        return process
-
-    def ask_record_ids(self, ids, single_coll=False):
-        """
-        Asks the user for a single or set of Record IDs.
-        
-        :param ids: A single or set of Record IDs with their collections.
-        :type  ids: str
-        """
-
-        if ids is None or ids == '':
-
-            if not self.eod.silent:
-                self.print_header("Enter Record Id(s)")
-
-                msg = "\nEnter a single or set of Record IDs. Include the " \
-                      "Collection ID at the start of IDs separated by a " \
-                      "pipe. Separate collection's Ids with a comma. " \
-                      f"(Ex: {self.eod.var_colour}" \
-                      f"RCMImageProducts:7625368|25654750" \
-                      f",NAPL:3736869{self.eod.reset_colour})\n"
-                if single_coll:
-                    msg = f"\nEnter a single or set of Record IDs with the " \
-                        f"Collection ID at the start of IDs separated by a " \
-                        f"pipe (Ex: {self.eod.var_colour}" \
-                        f"RCMImageProducts:7625368|25654750" \
-                        f"{self.eod.reset_colour})\n"
-                ids = self.get_input(msg, required=False)
-
-                process = self.eod.validate_record_ids(ids, single_coll)
-
-                if not process:
-                    err_msg = "Invalid entry for the Record Ids."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-
-        return ids
-
-    def ask_uuid(self, coll_id=None, uuid_val=None):
-        """
-        Asks the user for UUID(s) using click prompts for collection and items.
-        
-        :param coll_id: The collection ID if already set by the command-line.
-        :type  coll_id: str
-        :param uuid_val: The UUID(s) if already set by the command-line.
-        :type  uuid_val: str
-        
-        :return: A tuple containing (collection_id, uuid_list)
-        :rtype: tuple
-        """
-        
-        if coll_id is None or coll_id == '':
-            if not self.eod.silent:
-                self.print_header("Select Collection")
-                
-                # Get available collections
-                coll_lst = self.eod.get_collections(True)
-                coll_lst = sorted(coll_lst, key=lambda x: x['title'])
-                
-                # Display collections
-                coll_choices = {str(idx + 1): c for idx, c in enumerate(coll_lst)}
-                for idx, c in enumerate(coll_lst):
-                    print(f"{idx + 1}. {c['title']} ({c['id']})")
-                
-                # Use click to prompt for collection selection
-                choice_str = click.prompt("Enter the number of the collection")
-                
-                try:
-                    choice_idx = int(choice_str) - 1
-                    if choice_idx < 0 or choice_idx >= len(coll_lst):
-                        raise ValueError()
-                    coll_id = coll_lst[choice_idx]['id']
-                except (ValueError, IndexError):
-                    err_msg = "Invalid collection selection."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-            else:
-                err_msg = "Collection must be specified."
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-        
-        if uuid_val is None or uuid_val == '':
-            if not self.eod.silent:
-                self.print_header("Enter UUID(s)")
-                
-                # Use click to prompt for UUIDs
-                msg = "Enter one or more UUID(s), separated by commas"
-                uuid_val = click.prompt(msg)
-                
-                if not uuid_val or uuid_val.strip() == '':
-                    err_msg = "At least one UUID must be specified."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-            else:
-                err_msg = "UUID must be specified."
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-        
-        # Parse UUIDs (comma-separated)
-        uuids = [u.strip() for u in uuid_val.split(',')]
-        
-        return coll_id, uuids
-
-
-    def ask_st_images(self, ids):
-
-        """
-        Asks the user for Record IDs or Order Keys for SAR Toolbox orders.
-        
-        :param ids: A single or set of Record IDs with their collections.
-        :type  ids: str
-        """
-
-        if ids is None or ids == '':
-
-            if not self.eod.silent:
-                self.print_header("Enter Record Id(s) or Order Key(s)")
-                
-                msg = f"\nEnter a single or set of Record IDs or enter a " \
-                    f"single or set of Order Keys separated by a pipe. " \
-                    f"Include the Collection Id at the beginning of the set." \
-                    f" (Ex: {self.eod.var_colour}" \
-                    f"RCMImageProducts:7625368|25654750" \
-                    f"{self.eod.reset_colour} or {self.eod.var_colour}" \
-                    f"RCMImageProducts:RCM2_OK1373330_PK1530425_1_16M12_" \
-                    f"20210326_111202_HH_HV_GRD|RCM2_OK1373330_PK1524695_1_" \
-                    f"16M17_20210321_225956_HH_HV_GRD{self.eod.reset_colour})\n"
-                ids = self.get_input(msg, required=False)
-
-                process = self.eod.validate_st_images(ids)
-
-                if not process:
-                    err_msg = "Invalid entry for the Record Ids or Order Keys."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-
-        return ids
-
-    def ask_st(self):
-        """
-        Ask user for all SAR Toolbox information
-        """
-
-        def ask_param(param):
-
-            default = param.get_default(as_listidx=True, include_label=True)
-            # print(f"default: {default}")
-            if param.const_vals:
-                default_val = param.get_default(as_listidx=True)
-                default_str = param.get_default(as_listidx=True,
-                                                include_label=True)
-                labels = [c.get('label') for c in param.const_vals 
-                          if c.get('active')]
-                multiple = param.multiple
-                choice = ask_item(param.label, labels, 'param', 
-                                        multiple=multiple,
-                                        default=default_val,
-                                        def_msg=default_str)
-            else:
-                msg = f'Enter the "{param.get_label()}"'
-                choice = self.get_input(msg, required=False,
-                                        default=default)
-                      
-            val_check = param.set_value(choice)
-
-            if not val_check:
-                err_msg = f"An invalid value has been entered. The value has " \
-                            f"to be of type '{param.get_data_type()}'"
-                self.eod.print_msg(err_msg, heading='error')
-                self.logger.error(err_msg)
-                self.eod.exit_cli(1)
-
-            if param.const_vals:
-                choice = [labels[int(idx) - 1] for idx in choice]
-
-            if param.get_value():
-                sub_params = param.get_sub_param()
-                if param.data_type == 'bool' and param.get_value() == 'False':
-                    return None
-                if sub_params:
-                    for s_param in sub_params:
-                        if param.param_id == 'OutputPixSpacing':
-                            param_val = param.get_value(True)
-                            if param_val.lower() == 'meters' \
-                                    and s_param.param_id == 'OutputPixSpacingMeters':
-                                ask_param(s_param)
-                            elif param_val.lower() == 'degrees' \
-                                    and s_param.param_id == 'OutputPixSpacingDegrees':
-                                ask_param(s_param)
-                        else:
-                            ask_param(s_param)
-
-        def ask_item(item_name, item_list, item_type='runner', multiple=False,
-                     required=False, default=None, def_msg=None):
-            choice_strs = []
-            for idx, v in enumerate(item_list):
-                choice_strs.append(self.wrap_text(
-                                    f"{self.eod.var_colour}{idx + 1}" \
-                                    f"{self.eod.reset_colour}: {v}", 
-                                    sub_indent='     '))
-            choices = '\n'.join(choice_strs)
-
-            info_str = ""
-            if multiple:
-                info_str = " (separate each number with a comma)"
-
-            if item_type == 'runner':
-                msg = f'Which "{item_name}" would you like to run?'
-            elif item_type == 'product':
-                msg = f"Select the Output Layer options"
-            else:
-                msg = f'Available "{item_name}" options'
-            print(f'\n{msg}:\n\n{choices}')
-
-            if item_type == 'product':
-                msg = f'Please choose the Output Layer options{info_str}'
-            else:
-                msg = f'Please choose the "{item_name}"{info_str}'
-            choice = self.get_input(msg, required=required, default=default,
-                                    def_msg=def_msg)
-
-            if not multiple and required:
-                if not choice.isdigit():
-                    err_msg = "An invalid value has been entered."
-                    self.eod.print_msg(err_msg, heading='error')
-                    self.logger.error(err_msg)
-                    self.eod.exit_cli(1)
-            else:
-                if choice:
-                    choice = str(choice).split(',')
-
-                    for c in choice:
-                        if not c.isdigit() or int(c) > len(item_list) \
-                                or int(c) <= 0:
-                            err_msg = "An invalid value has been entered."
-                            self.eod.print_msg(err_msg, heading='error')
-                            self.logger.error(err_msg)
-                            self.eod.exit_cli(1)
-
-            return choice
-
-        self.print_header("Enter SAR Toolbox Information")
-
-        st = sar.SARToolbox(self.eod)
-        
-        ###############################
-        # Set the category
-        ###############################
-
-        # Ask for polarization to start
-        param = st.get_polarization_param()
-        def_msg = param.get_default(as_listidx=True, include_label=True)
-        default = param.get_default(as_listidx=True)
-        labels = [c.get('label') for c in param.const_vals]
-        multiple = param.multiple
-        choice_idx = ask_item(param.label, labels, 'param', 
-                                default=default, 
-                                multiple=multiple,
-                                def_msg=def_msg)
-        polarization = [labels[int(idx) - 1] for idx in choice_idx]
-        param.set_value(polarization)
-
-        cat_names = st.get_cat_names(True)
-        cat_indices = ask_item("Categories", cat_names, multiple=True, 
-                               required=True)
-        categories = st.set_category_runs(cat_indices)
-
-        for category in categories:
-            self.print_sub_header(f'Enter Methods for "{category.name}"')
-
-            ###############################
-            # Set the method
-            ###############################
-
-            methods = category.get_method_names(True)
-            method_indices = ask_item("Methods", methods, multiple=True, 
-                                      required=True)
-            methods = category.set_method_runs(method_indices)
-
-            for method in methods:
-
-                self.print_sub_header(f'Enter Arguments for '
-                                      f'"{category.name} - {method.name}"')
-
-                ###############################
-                # Ask for arguments
-                ###############################
-                params = method.get_parameters()
-                for param in params:
-                    ask_param(param)
-                    method.add_param_run(param)
-
-                ###############################
-                # Ask for products
-                ###############################
-
-                products = method.get_products()
-
-                if products:
-                    labels = [p.name for p in products]
-                    choice_idx = ask_item(param.label, labels, 'product', True)
-                    if choice_idx:
-                        choices = [products[int(c) - 1] for c in choice_idx]
-                        method.set_prod_runs(choices)
-
-                method.print_info()
-
-        msg = "If you'd like to save the SAR Toolbox JSON request, " \
-                "enter the file path"
-        save_st = self.get_input(msg, required=False)
-        st.set_output_fn(save_st)
-
-        return st
-
-    def build_syntax(self):
-        """
-        Builds the command-line syntax to print to the command prompt.
-        
-        :return: A string containing the command-line syntax for the script.
-        :rtype: str
-        """
-
-        click_ctx = click.get_current_context(silent=True)
-
-        flags = {}
-        if click_ctx is None:
-            return ''
-
-        cmd_params = click_ctx.to_info_dict()['command']['params']
-        for p in cmd_params:
-            flags[p['name']] = p['opts']
-
-
-        syntax_params = []
-        for p, pv in self.params.items():
-            if pv is None or pv == '':
-                continue
-            if p == 'session':
-                continue
-            if p == 'eodms_rapi':
-                continue
-
-            opts = flags.get(p)
-            if not opts:
-                continue
-
-            # Prefer short flag when available, otherwise use the only option.
-            if len(opts) > 1:
-                flag = opts[1]
-            else:
-                flag = opts[0]
-
-            if isinstance(pv, list):
-                if flag == '-d':
-                    pv = '-'.join(['"%s"' % i if isinstance(i, str) and i.find(' ') > -1 else str(i)
-                                   for i in pv])
-                else:
-                    pv = ','.join(['"%s"' % i if isinstance(i, str) and i.find(' ') > -1 else str(i)
-                                   for i in pv])
-
-            elif isinstance(pv, dict):
-                if flag == '-f':
-                    filt_lst = []
-                    for k, v_lst in pv.items():
-                        if isinstance(v_lst, str):
-                            v = v_lst.strip()
-                            if v != '':
-                                v = v.replace('"', '').replace("'", '')
-                                filt_lst.append(f"{k}.{v}")
-                        else:
-                            for v in v_lst:
-                                if v is None or v == '':
-                                    continue
-                                v = v.replace('"', '').replace("'", '')
-                                filt_lst.append(f"{k}.{v}")
-                    if len(filt_lst) == 0:
-                        continue
-                    pv = '"%s"' % ','.join(filt_lst)
-                elif flag == '-i':
-                    pv = f'"{pv}"'
-
-            elif isinstance(pv, bool):
-                if not pv:
-                    continue
-                else:
-                    pv = ''
-            else:
-                if isinstance(pv, str) and pv.find(' ') > -1:
-                    pv = f'"{pv}"'
-                elif isinstance(pv, str) and pv.find('|') > -1:
-                    pv = f'"{pv}"'
-                elif isinstance(pv, dict):
-                    pv = f'"{pv}"'
-
-            syntax_params.append(f'{flag} {pv}')
-
-        out_syntax = "python %s %s -s" % (os.path.realpath(__file__),
-                                          ' '.join(syntax_params))
-
-        return out_syntax
-
-    def get_input(self, msg, err_msg=None, required=True, options=None,
-                  default=None, def_msg=None, password=False):
-        """
-        Gets an input from the user for an argument.
-        
-        :param msg: The message used to prompt the user.
-        :type  msg: str
-        :param err_msg: The message to print when the user enters an invalid
-                input.
-        :type  err_msg: str
-        :param required: Determines if the argument is required.
-        :type  required: boolean
-        :param options: A list of available options for the user to choose from.
-        :type  options: list[str]
-        :param default: The default value if the user just hits enter.
-        :type  default: str
-        :param def_msg: If the default is None or an empty string, this 
-                        variable will tell the user what happens with no answer.
-        :type  def_msg: str
-        :param password: Determines if the argument is for password entry.
-        :type  password: boolean
-        
-        :return: The value entered by the user.
-        :rtype: str
-        """
-
-        opt_str = ''
-        if options is not None:
-            opt_join = '/'.join(options)
-            opt_str = f" ({self.eod.var_colour}{opt_join}" \
-                    f"{self.eod.reset_colour})"
-
-        def_str = ''
-        if default is not None:
-            def_str = f" {self.eod.def_colour}[{default}]" \
-                f"{self.eod.reset_colour}"
-        
-        if def_msg is not None:
-            def_str = f" {self.eod.def_colour}[{def_msg}]" \
-                f"{self.eod.reset_colour}"
-
-        output = f"\n{msg}{opt_str}{def_str}"
-        if msg.endswith('\n'):
-            msg_strp = msg.strip('\n')
-            output = f"\n{msg_strp}{opt_str}{def_str}\n"
-        
-        # Strip ANSI color codes for click.prompt display
-        output_display = output
-        output_display = output_display.replace(self.eod.var_colour, '')
-        output_display = output_display.replace(self.eod.def_colour, '')
-        output_display = output_display.replace(self.eod.reset_colour, '')
-        
-        wrapped_output = self.wrap_text(output_display)
-
-        # click.prompt requires a default to accept an empty response.
-        prompt_default = default
-        if not required and prompt_default is None:
-            prompt_default = ''
-        
+    s_intersect_list: List[Dict[str, Any]] = []
+    if aoi:
         try:
-            in_val = click.prompt(wrapped_output, default=prompt_default, 
-                                  hide_input=password, show_default=False)
-        except EOFError as error:
-            # Output expected EOFErrors.
-            self.logger.error(error)
-            eod_util.EodmsProcess().exit_cli(1)
+            s_intersect_list = parse_aoi_file(aoi)
+            click.echo(f"Loaded {len(s_intersect_list)} polygon(s) from AOI file")
+        except ValueError as exc:
+            raise click.ClickException(str(exc))
+    elif s_intersect:
+        s_intersect_list = [{"name": None, "wkt": s_intersect}]
 
-        if required and in_val == '':
-            # eod_util.EodmsProcess().print_support(True, err_msg)
-            if err_msg is None:
-                err_msg = "Parameter is required."
-            eod_util.EodmsProcess().print_msg(err_msg, heading='error')
-            self.logger.error(err_msg)
-            eod_util.EodmsProcess().exit_cli(1)
+    if not s_intersect_list:
+        s_intersect_list = [{"name": None, "wkt": None}]
 
-        if self.testing:
-            print(f"FOR TESTING - Value entered: {in_val}")
+    search_api = make_search(aaa_api, env)
+    items = search_api.search_multiple_geometries(
+        s_intersect_list=s_intersect_list,
+        collection=collection,
+        datetime_range=datetime_range,
+        bbox=bbox_list,
+        limit=limit,
+        filter_text=filter_text,
+    )
 
-        return in_val
-    
-    def add_arrow(self):
-        """
-        Adds an arrow (->>) to the output
+    if not items:
+        click.echo("No items found.")
+        return
 
-        :return: A string containing a coloured arrow.
-        :rtype: str
-        """
-        arrow = f"{self.eod.arrow_colour}->>{self.eod.reset_colour}"
+    click.echo(f"Found {len(items)} item(s).")
 
-        return arrow
-    
-    def wrap_text(self, in_str, width=105, sub_indent='  ', init_indent=''):
-        """
-        Wraps a given text to a certain width.
+    if output:
+        save_items_geojson(items, output)
+        return
 
-        :param in_str: The input string to wrap.
-        :type  in_str: str
-        :param width: The width of the text.
-        :type  width: int
+    preview_uuids: List[str] = []
+    for item in items:
+        item_uuid = _extract_item_uuid(item)
+        if item_uuid:
+            preview_uuids.append(item_uuid)
+        if len(preview_uuids) == 5:
+            break
 
-        :return: The input string now wrapped.
-        :rtype: str
-        """
+    if preview_uuids:
+        click.echo("First 5 UUID(s):")
+        for item_uuid in preview_uuids:
+            click.echo(item_uuid)
+    else:
+        click.echo("No UUID values found in search results.")
 
-        out_str = textwrap.fill(in_str, width=width, 
-                                replace_whitespace=False, 
-                                initial_indent=init_indent,
-                                subsequent_indent=sub_indent)
-        return out_str
-    
-    def print_header(self, header):
-        """
-        Prints the header for input
-        """
 
-        print(f"{self.eod.head_colour}\n--------------{header}--------------" \
-              f"{self.eod.reset_colour}")
+@cli.command("process")
+@click.option("--username", "-u", required=False, help="EODMS username.")
+@click.option("--password", "-p", required=False, help="EODMS password.")
+@click.option("--env", "-e", required=False, default="prod",
+              help='Environment (default: "prod").')
+@click.option("--process_id", "-pi", required=False, default=None,
+              help="Processing service ID.")
+@click.option("--list/--no-list", "list_processes", default=True,
+              help="List available processes (default behavior).")
+@click.option("--describe", "describe", is_flag=True,
+              help="Print process input structure and sample payload.")
+@click.option("--submit", is_flag=True,
+              help="Submit a processing job (requires auth).")
+@click.option("--inputs_json", "--input_json", required=False, default=None,
+              help="JSON string or path to JSON file for submit inputs.")
+@click.option("--outputs_json", required=False, default=None,
+              help="JSON string or path to JSON file for generic submit outputs.")
+@click.option("--mode", required=False, default="async",
+              help="Execution mode for generic submit (default: async).")
+@click.option("--job_id", "-j", required=False, default=None,
+              help="Existing job ID to check, poll, retrieve results, or download outputs.")
+@click.option("--wait", is_flag=True,
+              help="Poll job status until terminal state.")
+@click.option("--interval", required=False, default=30, type=int,
+              help="Polling interval seconds for --wait (default: 30).")
+@click.option("--timeout", required=False, default=600, type=int,
+              help="Polling timeout seconds for --wait (default: 600).")
+@click.option("--show_results", is_flag=True,
+              help="Print job results JSON.")
+@click.option("--download_dir", "-dl", required=False, default=None,
+              help="Download all job result files to this folder.")
+@click.option("--skip_existing/--no-skip_existing", default=True,
+              help="Skip existing local files when downloading results (default: enabled).")
+@click.option("--output", "-o", required=False, default=None,
+              help="Write JSON response (process details or submit response) to file.")
+@handle_service_errors
+def order_st_cmd(
+    username: Optional[str],
+    password: Optional[str],
+    env: str,
+    process_id: Optional[str],
+    list_processes: bool,
+    describe: bool,
+    submit: bool,
+    inputs_json: Optional[str],
+    outputs_json: Optional[str],
+    mode: str,
+    job_id: Optional[str],
+    wait: bool,
+    interval: int,
+    timeout: int,
+    show_results: bool,
+    download_dir: Optional[str],
+    skip_existing: bool,
+    output: Optional[str],
+):
+    """Processing Service for RADARDAT data; Level-1, SAR Toolbox, and ARD."""
 
-    def print_sub_header(self, header):
-        """
-        Prints the header for input
-        """
+    username, password = resolve_credentials(username, password)
 
-        print(f"{self.eod.head_colour}\n=== {header} ===" \
-              f"{self.eod.reset_colour}")
+    aaa_api = make_aaa(username, password, env)
+    proc_api = make_processes(aaa_api, env)
 
-    def print_syntax(self):
-        """
-        Prints the command-line syntax for the script.
-        """
+    if list_processes and not submit and not describe and not job_id:
+        processes_json = proc_api.list_processes()
+        _print_process_summary(processes_json)
+        return
 
-        print("\nUse this command-line syntax to run the same parameters:")
-        self.cli_syntax = self.build_syntax()
-        print(f"{self.eod.path_colour}{self.cli_syntax}{self.eod.reset_colour}")
-        self.logger.info(f"Command-line Syntax: {self.cli_syntax}")
+    if describe:
+        if not process_id:
+            raise click.UsageError("--process_id is required with --describe")
 
-    def is_json_dict(self, in_str):
+        if str(process_id).strip() == "SAR_Toolbox":
+            schema_url = "https://eodms-sgdot.nrcan-rncan.gc.ca/schemas/st/sar-toolbox-schema.json"
+            try:
+                with urlopen(schema_url) as resp:
+                    body = resp.read().decode("utf-8")
+                schema_json = json.loads(body)
+            except (URLError, ValueError) as exc:
+                raise click.ClickException(f"Failed to fetch SAR Toolbox schema: {exc}")
+
+            click.echo(json.dumps(schema_json, indent=4))
+            if output:
+                with open(output, "w", encoding="utf-8") as out_f:
+                    json.dump(schema_json, out_f, indent=2)
+                click.echo(f"Saved SAR Toolbox schema to {output}")
+            return
+
+        process_json = proc_api.get_process(process_id)
+        click.echo(json.dumps(process_json.get("inputs", {}), indent=4))
+
+        sample_payload = _build_sample_payload(process_id, process_json)
+        click.echo("\nSample execution payload:")
+        click.echo(json.dumps(sample_payload, indent=4))
+
+        if output:
+            with open(output, "w", encoding="utf-8") as out_f:
+                json.dump(process_json, out_f, indent=2)
+            click.echo(f"Saved process description to {output}")
+        return
+
+    submitted_job_id = None
+    if submit:
+        if not process_id:
+            raise click.UsageError("--process_id is required with --submit")
+        if aaa_api is None:
+            raise click.UsageError("--username and --password are required with --submit")
+
+        loaded_inputs = _load_json_input(inputs_json, "inputs")
+        if str(process_id).strip() == "SAR_Toolbox":
+            if loaded_inputs is None:
+                raise click.UsageError("--input_json (or --inputs_json) is required with --submit for SAR_Toolbox")
+
+            sar_request = loaded_inputs
+            if isinstance(loaded_inputs, dict) and "inputs" in loaded_inputs and isinstance(loaded_inputs.get("inputs"), dict):
+                nested_inputs = loaded_inputs.get("inputs")
+                if isinstance(nested_inputs, dict) and "items" in nested_inputs:
+                    sar_request = nested_inputs
+
+            if not isinstance(sar_request, dict) or not isinstance(sar_request.get("items"), list):
+                raise click.UsageError(
+                    "SAR_Toolbox submit expects a JSON request containing an 'items' list "
+                    "(same structure used by legacy order_st / st_request)."
+                )
+
+            rapi_api = EODMSRAPI(username, password)
+            submit_json = _safe_rapi_call(rapi_api.order_json, sar_request)
+            click.echo(json.dumps(submit_json, indent=2))
+
+            if output:
+                with open(output, "w", encoding="utf-8") as out_f:
+                    json.dump(submit_json, out_f, indent=2)
+                click.echo(f"Saved submission response to {output}")
+
+            if download_dir:
+                order_id = _extract_first_order_id(submit_json)
+                if not order_id:
+                    raise click.ClickException("No orderId found in SAR_Toolbox submit response; cannot download.")
+                order_payload = _safe_rapi_call(rapi_api.get_order, order_id)
+                order_items = _safe_rapi_call(rapi_api.collect_order_items, order_payload)
+                _download_rapi_items(rapi_api, order_items, download_dir)
+
+            return
+
+        outputs = _load_json_input(outputs_json, "outputs")
+        if loaded_inputs is None:
+            raise click.UsageError("--inputs_json is required with --submit")
+
+        request_mode = mode
+        if isinstance(loaded_inputs, dict) and "inputs" in loaded_inputs:
+            inputs = loaded_inputs.get("inputs")
+            if outputs is None and isinstance(loaded_inputs.get("outputs"), dict):
+                outputs = loaded_inputs.get("outputs")
+            if isinstance(loaded_inputs.get("mode"), str) and loaded_inputs.get("mode").strip():
+                request_mode = loaded_inputs.get("mode").strip()
+        else:
+            inputs = loaded_inputs
+
+        if not isinstance(inputs, dict):
+            raise click.UsageError("Resolved submit inputs must be a JSON object.")
+
+        submit_json = proc_api.submit_process(
+            process_id=process_id,
+            inputs=inputs,
+            outputs=outputs,
+            mode=request_mode,
+        )
+
+        click.echo(json.dumps(submit_json, indent=2))
+        submitted_job_id = submit_json.get("jobID")
+        click.echo(f"Submitted jobID: {submitted_job_id}")
+
+        if output:
+            with open(output, "w", encoding="utf-8") as out_f:
+                json.dump(submit_json, out_f, indent=2)
+            click.echo(f"Saved submission response to {output}")
+
+    target_job_id = job_id or submitted_job_id
+
+    if wait:
+        if not target_job_id:
+            raise click.UsageError("A job ID is required for --wait (provide --job_id or --submit)")
+        status_json = proc_api.poll_job_status(target_job_id, interval=interval, timeout=timeout)
+        click.echo(json.dumps(status_json, indent=2))
+    elif target_job_id and not show_results and not download_dir:
+        status_json = proc_api.get_job_status(target_job_id)
+        click.echo(json.dumps(status_json, indent=2))
+
+    if show_results:
+        if not target_job_id:
+            raise click.UsageError("A job ID is required for --show_results")
+        results_json = proc_api.get_job_results(target_job_id)
+        click.echo(json.dumps(results_json, indent=2))
+
+    if download_dir:
+        if not target_job_id:
+            raise click.UsageError("A job ID is required for --download_dir")
+        downloaded = proc_api.download_job_results(
+            job_id=target_job_id,
+            out_dir=os.path.abspath(download_dir),
+            skip_existing=skip_existing,
+        )
+        click.echo(json.dumps({"jobID": target_job_id, "downloaded_files": downloaded}, indent=2))
+
+
+@cli.command("download")
+@click.option("--username", "-u", required=False, help="EODMS username.")
+@click.option("--password", "-p", required=False, help="EODMS password.")
+@click.option("--uuid", required=False, default=None,
+              help="Download UUID directly via STAC assets (public collections) or DDS.")
+@click.option("--input", "input_file", required=False, default=None, type=click.Path(exists=True),
+              help="Input file for download items (GeoJSON, JSON, or JSONL retry records).")
+@click.option("--collection", "-c", required=False, default=None,
+              help="Collection for --uuid download.")
+@click.option("--env", "-e", required=False, default="prod",
+              help='Environment for DDS download (default: "prod").')
+@click.option("--order-items", required=False, default=None,
+              help="Legacy selector syntax: order:id1,id2|item:id3,id4")
+@click.option("--limit", "-l", required=False, default=100, type=int,
+              help="Maximum AVAILABLE_FOR_DOWNLOAD orders to retrieve.")
+@click.option("--dtstart", required=False, default=None,
+              help="Optional start datetime filter for order retrieval.")
+@click.option("--dtend", required=False, default=None,
+              help="Optional end datetime filter for order retrieval.")
+@click.option("--list", "list_orders", is_flag=True,
+              help="List orders in the last 30 days and exit (no download).")
+@click.option("--order-status", required=False, default=None,
+              help="Optional status filter for --list (example: AVAILABLE_FOR_DOWNLOAD).")
+@click.option("--download-available", is_flag=True,
+              help="Download AVAILABLE_FOR_DOWNLOAD order items (required for default bulk download mode).")
+@click.option("--dl_dir", "download_dir", required=False, default=".\\downloads",
+              help="Destination directory for downloads.")
+@click.option("--download-dir", "download_dir", required=False,
+              help="Destination directory for downloads.")
+@click.option("--dds-retry-file", required=False, default=DEFAULT_DDS_RETRY_FILE,
+              help="File used to append DDS ItemRestoring records for retry input.")
+@click.pass_context
+@handle_service_errors
+def download_available_cmd(
+    ctx: click.Context,
+    username: Optional[str],
+    password: Optional[str],
+    uuid: Optional[str],
+    input_file: Optional[str],
+    collection: Optional[str],
+    env: str,
+    order_items: Optional[str],
+    limit: int,
+    dtstart: Optional[str],
+    dtend: Optional[str],
+    list_orders: bool,
+    order_status: Optional[str],
+    download_available: bool,
+    download_dir: str,
+    dds_retry_file: str,
+):
+    """Download by UUID (public STAC assets or DDS) and legacy RAPI order-item downloads."""
+
+    username, password = resolve_credentials(username, password)
+    dds_backoff_seconds = _load_dds_backoff_interval()
+    dds_concurrent_downloads = _load_dds_concurrent_downloads()
+
+    # If explicitly provided, allow --dds-retry-file to act as --input for retry runs.
+    dds_retry_file_provided = False
+    if hasattr(ctx, "get_parameter_source"):
         try:
-            in_str = in_str.replace("'", '"')
-            res = json.loads(in_str)
-            return res
-        except json.JSONDecodeError:
-            return False
-
-    def prompt(self):
-        """
-        Prompts the user for the input options.
-        """
-
-        username = self.params.get('username')
-        password = self.params.get('password')
-        input_val = self.params.get('input_val')
-        collections = self.params.get('collections')
-        process = self.params.get('process')
-        filters = self.params.get('filters')
-        dates = self.params.get('dates')
-        maximum = self.params.get('maximum')
-        output = self.params.get('output')
-        aws = self.params.get('aws')
-        overlap = self.params.get('overlap')
-        orderitems = self.params.get('orderitems')
-        no_order = self.params.get('no_order')
-        downloads = self.params.get('downloads')
-        st_request = self.params.get('st_request')
-        uuid = self.params.get('uuid')
-        collection = self.params.get('collection')
-        silent = self.params.get('silent')
-        version = self.params.get('version')
-
-        if version:
-            print(f"{__title__}: Version {__version__}")
-            self.eod.exit_cli()
-
-        self.eod.set_silence(silent)
-
-        new_user = False
-        new_pass = False
-
-        if username is None or password is None:
-            self.print_header("Enter EODMS Credentials")
-
-        if username is None or username == '':
-
-            username = self.config_util.get('Credentials', 'username')
-
-            if username == '':
-                msg = "Enter the username for authentication"
-                err_msg = "A username is required to order images."
-                username = self.get_input(msg, err_msg)
-                new_user = True
-            else:
-                print(f"\nUsing the username set in the " 
-                      f"'{self.eod.path_colour}"
-                      f"{self.config_util.get_filename()}" 
-                      f"{self.eod.reset_colour}' file...")
-
-        if password is None or password == '':
-
-            password = self.config_util.get('Credentials', 'password')
-
-            if password == '':
-                msg = 'Enter the password for authentication'
-                err_msg = "A password is required to order images."
-                password = self.get_input(msg, err_msg, password=True)
-                new_pass = True
-            else:
-                try:
-                    password = base64.b64decode(password).decode("utf-8")
-                except binascii.Error as err:
-                    password = base64.b64decode(password +
-                                                "========").decode("utf-8")
-                print(f"Using the password set in the " 
-                      f"'{self.eod.path_colour}"
-                      f"{self.config_util.get_filename()}" 
-                      f"{self.eod.reset_colour}' file...\n")
-
-        if new_user or new_pass:
-            suggestion = ''
-            if self.eod.silent:
-                suggestion = " (it is best to store the credentials if " \
-                             "you'd like to run the script in silent mode)"
-            msg = f"Would you like to store the credentials for a future " \
-                f"session{suggestion}? (y/n)"
-            answer = self.get_input(msg, required=False, default='n')
-            if answer.lower().find('y') > -1:
-                self.config_util.set('Credentials', 'username', username)
-                pass_enc = base64.b64encode(password.encode("utf-8")).decode(
-                    "utf-8")
-                self.config_util.set('Credentials', 'password', str(pass_enc))
-
-                self.config_util.write()
-
-        eodms_domain = self.eod.eodms_domain
-        if eodms_domain and eodms_domain.find('staging'):
-            print("\n**** RUNNING IN STAGING ENVIRONMENT ****\n")
-
-        self.eod.create_session(username, password)
-
-        self.params = {'collections': collections,
-                       'dates': dates,
-                       'input_val': input_val,
-                       'maximum': maximum,
-                       'process': process,
-                       'downloads': downloads,
-                       'uuid': uuid,
-                       'collection': collection}
-
-        print()
-        coll_dict = self.eod.get_collections(True)
-
-        self.eod.check_error(coll_dict)
-
-        print("\n(For more information on the following prompts, please refer"
-              " to the README file.)")
-
-        #########################################
-        # Get the type of process
-        #########################################
-
-        if process is None or process == '':
-            self.process = self.ask_process()
-        else:
-            self.process = process
-
-        if self.process == 'search_only':
-            msg = "The 'search_only' process is no longer available. " \
-                  "In future, please use the flags '--no_order' or '-nord' " \
-                  "to suppress ordering and downloading. Script will " \
-                  "perform a search without ordering or downloading."
-            self.eod.print_msg(msg, heading='note')
-            self.logger.warning(msg)
-            self.process = 'full'
-            no_order = True
-
-        proc_num = list(proc_choices.keys()).index(self.process) + 1
-        print(self.eod.title_colour)
-        print("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
-        print(f" Running Process "
-              f"{proc_num}: "
-              f"{proc_choices[self.process]['name']}")
-        print("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
-        print(self.eod.reset_colour)
-
-        self.params['process'] = self.process
-
-        if self.process == 'download_only':
-            self.eod.print_msg("The process 'download_only' is now named "
-                  "'download_results'. Please update any command-line "
-                  "syntaxes.", heading='note')
-            self.process = 'download_results'
-
-        if self.process == 'full':
-
-            self.logger.info("Searching, ordering and downloading images "
-                             "using an AOI.")
-
-            # Get the collection(s)
-            coll = self.ask_collection(collections, coll_lst=coll_dict)
-            self.params['collections'] = coll
-
-            # If Radarsat-1, ask user if they want to download from AWS
-            if 'Radarsat1' in coll:
-                aws = self.ask_aws(aws)
-                self.params['aws'] = aws
-
-            # Get the AOI file
-            inputs = self.ask_aoi(input_val)
-            self.params['input_val'] = inputs
-
-            # If an AOI is specified, ask for a minimum overlap percentage
-            if inputs is not None:
-                overlap = self.ask_overlap(overlap)
-                self.params['overlap'] = overlap
-
-            # Get the filter(s)
-            filt_dict = self.ask_filter(filters)
-            self.params['filters'] = filt_dict
-
-            # Get the date(s)
-            dates = self.ask_dates(dates)
-            self.params['dates'] = dates
-
-            # Ask user if they'd like to order and download
-            no_order = self.ask_order(no_order)
-            self.params['no_order'] = no_order
-
-            # Get the maximum(s)
-            maximum = self.ask_maximum(maximum)
-            self.params['maximum'] = maximum
-
-            # Get the output filename
-            output = self.ask_output(output)
-            self.params['output'] = output
-
-            # Print command-line syntax for future processes
-            self.print_syntax()
-
-            self.eod.search_order_download(self.params)
-
-        elif self.process == 'order_csv':
-
-            self.logger.info("Ordering and downloading images using results "
-                             "from a CSV file.")
-
-            #########################################
-            # Get the CSV file
-            #########################################
-
-            msg = "Enter the full path of the CSV file exported " \
-                  "from the EODMS UI website"
-            inputs = self.ask_input_file(input_val, msg)
-            self.params['input_val'] = inputs
-
-            # If Radarsat-1, ask user if they want to download from AWS
-            if os.path.exists(inputs):
-                lines = open(inputs, 'r').read()
-                if lines.lower().find('radarsat-1') > -1:
-                    aws = self.ask_aws(aws)
-                    self.params['aws'] = aws
-
-            # Ask user if they'd like to order and download
-            no_order = self.ask_order(no_order)
-            self.params['no_order'] = no_order
-
-            # Get the maximum(s)
-            maximum = self.ask_maximum(maximum)
-            self.params['maximum'] = maximum
-
-            # Get the output filename
-            output = self.ask_output(output)
-            self.params['output'] = output
-
-            # Print command-line syntax for future processes
-            self.print_syntax()
-
-            # Run the order_csv process
-            self.eod.order_csv(self.params)
-
-        elif self.process == 'download_available':
-            self.logger.info("Downloading existing order items with status"
-                             "AVAILABLE_FOR_DOWNLOAD.")
-
-            orderitems = self.ask_orderitems(orderitems)
-            self.params['orderitems'] = orderitems
-
-            if orderitems is None or orderitems == '':
-                # Get the maximum(s)
-                maximum = self.ask_maximum(maximum, 'download')
-                self.params['maximum'] = maximum
-
-            # Get the output geospatial filename
-            output = self.ask_output(output)
-            self.params['output'] = output
-
-            # Print command-line syntax for future processes
-            self.print_syntax()
-
-            # Run the download_available process
-            self.eod.download_available(self.params)
-
-        elif self.process == 'uuid':
-            # Order and download a single or set of images using UUIDs
-
-            self.logger.info("Ordering and downloading images using UUIDs")
-
-            coll_id = self.params.get('collection')
-            uuid_val = self.params.get('uuid')
-            
-            coll_id, uuids = self.ask_uuid(coll_id, uuid_val)
-            self.params['collection'] = coll_id
-            self.params['uuid'] = ','.join(uuids)
-
-            # Ask user if they'd like to order and download
-            no_order = self.ask_order(no_order)
-            self.params['no_order'] = no_order
-
-            # Get the output filename
-            output = self.ask_output(output)
-            self.params['output'] = output
-
-            # Print command-line syntax for future processes
-            self.print_syntax()
-
-            # Run the order by uuid process
-            self.eod.order_uuids(self.params)
-
-        elif self.process == 'download_restored_items':
-            self.logger.info("Downloading previous requests with status ItemsRestoring")
-
-            # Get the CSV file
-            msg = "Enter the full path of the JSON ItemsRestore file from a " \
-                  "previous session"
-            inputs = self.ask_input_file(input_val, msg, file_type='json')
-            self.params['input_val'] = inputs
-
-            # Run the order_csv process
-            self.eod.download_restored_items(self.params)
-
-        elif self.process == 'order_st':
-            self.logger.info("Ordering an image from the SAR Toolbox")
-
-            if st_request:
-                self.params['st_request'] = st_request
-                sar_tb = sar.SARToolbox(self.eod, out_fn=st_request)
-                sar_tb.ingest_request()
-            else:
-                inputs = self.ask_st_images(input_val)
-
-                print(f"inputs: {inputs}")
-
-                inputs_dict = self.is_json_dict(inputs)
-                if inputs_dict:
-                    self.params['input_val'] = inputs_dict
-                else:
-                    coll_id, ids = inputs.split(':')
-                    # If Order Keys are entered, check if they exist
-                    if inputs.find("_") > -1:
-                        ord_keys = ids.split('|')
-                        rec_ids = self.eod.get_record_ids(coll_id, ord_keys)
-
-                        if len(rec_ids) == 0:
-                            err_msg = f"No images could be found with Order Keys: "\
-                                    f"{', '.join(ord_keys)}."
-                            self.eod.logger.error(err_msg)
-                            # self.print_support(err_msg)
-                            self.eod.print_msg(err_msg, heading='error')
-                            self.eod.exit_cli(1)
+            src = ctx.get_parameter_source("dds_retry_file")
+            param_source = getattr(click.core, "ParameterSource", None)
+            if param_source is not None and src == param_source.COMMANDLINE:
+                dds_retry_file_provided = True
+        except Exception:
+            dds_retry_file_provided = False
+
+    if not uuid and not input_file and dds_retry_file_provided:
+        retry_input_path = os.path.abspath(dds_retry_file)
+        if not os.path.exists(retry_input_path):
+            raise click.ClickException(
+                f"DDS retry file not found: {retry_input_path}"
+            )
+        click.echo(f"Using DDS retry input file: {retry_input_path}")
+        input_file = retry_input_path
+
+    if uuid and input_file:
+        raise click.ClickException("Use either --uuid or --input, not both.")
+
+    if input_file:
+        features = _load_download_items_from_geojson(input_file)
+        if not features:
+            click.echo("No items found in input file.")
+            return
+
+        aaa_api = make_aaa(username, password, env)
+        search_api = None
+        dds_api = None
+        dds_work_items: List[Tuple[str, str]] = []
+
+        total_features = len(features)
+        processed_count = 0
+        skipped_count = 0
+        asset_download_count = 0
+        dds_download_count = 0
+        dds_retry_logged_count = 0
+        dds_no_download_count = 0
+
+        click.echo(f"Found {total_features} item(s) in input file.")
+        for feature in features:
+            item_uuid = _extract_item_uuid(feature)
+            item_collection = collection or feature.get("collection")
+
+            if not item_collection and isinstance(feature.get("properties"), dict):
+                item_collection = feature.get("properties", {}).get("collection")
+
+            if not item_uuid:
+                click.echo("Skipping item with no UUID/id field.")
+                skipped_count += 1
+                continue
+
+            if not item_collection:
+                click.echo(f"Skipping uuid={item_uuid}: no collection on feature and no --collection override.")
+                skipped_count += 1
+                continue
+
+            if _is_public_asset_collection(str(item_collection)):
+                if search_api is None:
+                    search_api = make_search(aaa_api, env)
+                click.echo(f"Downloading public assets: collection={item_collection}, uuid={item_uuid}")
+                downloaded_assets = download_public_stac_assets(search_api, str(item_collection), item_uuid, download_dir)
+                asset_download_count += downloaded_assets
+                processed_count += 1
+                continue
+
+            if not username or not password:
+                click.echo(
+                    f"Skipping uuid={item_uuid} in collection={item_collection}: credentials required for DDS."
+                )
+                skipped_count += 1
+                continue
+
+            if dds_api is None:
+                dds_api = make_dds(aaa_api, env)
+            dds_work_items.append((str(item_collection), item_uuid))
+
+        if dds_work_items:
+            click.echo(
+                f"Downloading {len(dds_work_items)} DDS item(s) with "
+                f"up to {dds_concurrent_downloads} concurrent worker(s)..."
+            )
+
+            def _dds_worker(work_item: Tuple[str, str]) -> Optional[Dict[str, Any]]:
+                work_collection, work_uuid = work_item
+                click.echo(f"Downloading DDS item: collection={work_collection}, uuid={work_uuid}")
+                return download_dds_item(
+                    dds_api,
+                    work_collection,
+                    work_uuid,
+                    download_dir,
+                    queued_backoff_seconds=dds_backoff_seconds,
+                    retry_file=dds_retry_file,
+                )
+
+            with ThreadPoolExecutor(max_workers=dds_concurrent_downloads) as executor:
+                future_map = {
+                    executor.submit(_dds_worker, work_item): work_item
+                    for work_item in dds_work_items
+                }
+
+                for future in as_completed(future_map):
+                    work_collection, work_uuid = future_map[future]
+                    try:
+                        item_info = future.result()
+                    except Exception as exc:
+                        click.echo(
+                            f"DDS download failed: collection={work_collection}, "
+                            f"uuid={work_uuid}, error={exc}"
+                        )
+                        dds_no_download_count += 1
+                        continue
+
+                    if isinstance(item_info, dict) and item_info.get("_dds_retry_logged"):
+                        dds_retry_logged_count += 1
+                    elif item_info is not None:
+                        dds_download_count += 1
                     else:
-                        rec_ids = ids.split('|')
+                        dds_no_download_count += 1
 
-                    print(f"\nSubmitting images with Record Ids: " \
-                        f"{', '.join(rec_ids)}")
+            processed_count += len(dds_work_items)
 
-                    self.params['input_val'] = {"collection_id": coll_id, 
-                                                "record_ids": rec_ids}
+        click.echo(
+            "Input download summary: "
+            f"processed_features={processed_count}, skipped_features={skipped_count}, "
+            f"public_assets_downloaded={asset_download_count}, dds_items_downloaded={dds_download_count}, "
+            f"dds_retry_logged={dds_retry_logged_count}, dds_not_downloaded={dds_no_download_count}"
+        )
+        return
 
-                sar_tb = self.ask_st()
+    if uuid:
+        if not collection:
+            raise click.ClickException("--collection is required when using --uuid.")
 
-            # self.params['st_request'] = sar_tb.out_fn
-            # Print command-line syntax for future processes
-            self.print_syntax()
+        if _is_public_asset_collection(collection):
+            aaa_api = make_aaa(username, password, env)
+            search_api = make_search(aaa_api, env)
+            click.echo(f"Downloading all available STAC assets for UUID: {uuid}")
+            downloaded_count = download_public_stac_assets(search_api, collection, uuid, download_dir)
+            click.echo(f"Downloaded {downloaded_count} asset(s).")
+            return
 
-            # Run the order_csv process
-            self.eod.order_st(sar_tb, self.params)
+        if not username or not password:
+            raise click.ClickException(
+                "--username and --password are required for DDS UUID downloads "
+                "(non-public collections)."
+            )
 
+        aaa_api = make_aaa(username, password, env)
+        dds_api = make_dds(aaa_api, env)
+        click.echo(f"Downloading UUID: {uuid}")
+        download_dds_item(
+            dds_api,
+            collection,
+            uuid,
+            download_dir,
+            queued_backoff_seconds=dds_backoff_seconds,
+            retry_file=dds_retry_file,
+        )
+        return
+
+    if not username or not password:
+        raise click.ClickException(
+            "--username and --password are required for order listing/downloading."
+        )
+
+    if not order_items and not list_orders and not download_available:
+        raise click.ClickException(
+            "Use --download-available to download AVAILABLE_FOR_DOWNLOAD items, "
+            "or use --list to list only."
+        )
+
+    rapi_api = EODMSRAPI(username, password)
+
+    if list_orders:
+        if dtstart and dtend:
+            list_dtstart = _parse_iso_datetime(dtstart)
+            list_dtend = _parse_iso_datetime(dtend)
+            if list_dtstart is None or list_dtend is None:
+                raise click.ClickException("--dtstart and --dtend must be ISO datetime strings (example: 2026-05-21T00:00:00Z).")
+            list_dtstart_label = str(dtstart)
+            list_dtend_label = str(dtend)
         else:
-            self.eod.print_msg("That is not a valid process type.", 
-                               heading='error')
-            self.logger.error("An invalid parameter was entered during "
-                              "the prompt.")
-            self.eod.exit_cli(1)
+            now_utc = datetime.now(timezone.utc)
+            list_dtend = now_utc
+            list_dtstart = now_utc - timedelta(days=30)
+            list_dtstart_label = _format_rapi_datetime(list_dtstart)
+            list_dtend_label = _format_rapi_datetime(list_dtend)
 
-#------------------------------------------------------------------------------
+        click.echo(f"Getting list of orders from {list_dtstart_label} to {list_dtend_label}...")
+        list_params: Dict[str, Any] = {
+            "maxOrders": limit,
+            "dtstart": _to_rapi_orders_dt_string(list_dtstart),
+            "dtend": _to_rapi_orders_dt_string(list_dtend),
+            "status": order_status.upper() if order_status else None,
+            "collection": collection,
+            "env": env,
+        }
+        click.echo(f"List parameters: {json.dumps(list_params, indent=2)}")
+        request_url = _build_rapi_orders_url(
+            rapi_api,
+            max_orders=limit,
+            dtstart=list_dtstart,
+            dtend=list_dtend,
+            status=order_status,
+        )
+        click.echo(f"RAPI GET URL: {request_url}")
+        payload = _safe_rapi_call(
+            rapi_api.get_order_summaries,
+            max_orders=limit,
+            dtstart=list_dtstart,
+            dtend=list_dtend,
+            status=order_status,
+        )
+        orders = payload or []
 
-output_help = "The output file path containing the results. The output " \
-                "parameter can be: "
-output_help += "None (empty): No output will be created | "
-output_help += "CSV: For a list of results (not geospatial) | "
-output_help += "GeoJSON: The output will be in the GeoJSON format (use " \
-                "extension .geojson or .json) | "
-output_help += "KML: The output will be in KML format (use extension .kml) " \
-                "(requires GDAL Python package) | "
-output_help += "GML: The output will be in GML format (use extension .gml) " \
-                "(requires GDAL Python package) | "
-output_help += "Shapefile: The output will be ESRI Shapefile (requires GDAL " \
-                "Python package) (use extension .shp)"
+        if not orders:
+            click.echo("No orders found.")
+            return
 
-abs_path = os.path.abspath(__file__)
+        click.echo(f"\nFound {len(orders)} order(s).")
+        for order in orders:
+            flat_order = {
+                "order_id": order.get("order_id") or "N/A",
+                "status": order.get("status") or "N/A",
+                "submitted": order.get("submitted") or "N/A",
+                "updated": order.get("updated") or "N/A",
+                "priority": order.get("priority") or "N/A",
+                "items": order.get("items") if order.get("items") is not None else 0,
+                "record_ids": order.get("record_ids") or [],
+                "collections": order.get("collections") or [],
+                "name": order.get("names") or [],
+                "destinations": order.get("destinations") or [],
+            }
+            click.echo(json.dumps(flat_order, indent=2, ensure_ascii=True))
+        return
 
-def get_configuration_values(config_util, download_path):
+    downloadable_items: List[Dict[str, Any]] = []
+    filtered_items: List[Dict[str, Any]] = []
 
-    config_params = {}
+    if order_items:
+        order_ids, item_ids = parse_legacy_order_items(order_items)
 
-    # Set the various paths
-    if download_path is None or download_path == '':
-        download_path = config_util.get('Paths', 'downloads')
+        for order_id in order_ids:
+            payload = _safe_rapi_call(rapi_api.get_order, order_id)
+            downloadable_items.extend(_safe_rapi_call(rapi_api.collect_order_items, payload))
 
-        if download_path == '':
-            download_path = os.path.join(os.path.dirname(abs_path),
-                                         'downloads')
-        elif not os.path.isabs(download_path):
-            download_path = os.path.join(os.path.dirname(abs_path),
-                                         download_path)
-    config_params['download_path'] = download_path
+        for item_id in item_ids:
+            payload = _safe_rapi_call(rapi_api.get_order_item, item_id)
+            downloadable_items.extend(_safe_rapi_call(rapi_api.collect_order_items, payload))
 
-    res_path = config_util.get('Paths', 'results')
-    if res_path == '':
-        res_path = os.path.join(os.path.dirname(abs_path), 'results')
-    elif not os.path.isabs(res_path):
-        res_path = os.path.join(os.path.dirname(abs_path),
-                                res_path)
-    config_params['res_path'] = res_path
+        # Explicit order/item selectors can include non-downloadable states.
+        for item in downloadable_items:
+            status = str(item.get("status", "")).upper()
+            if not status or status == "AVAILABLE_FOR_DOWNLOAD":
+                filtered_items.append(item)
+    else:
+        bulk_dtstart = _parse_iso_datetime(dtstart) if dtstart else None
+        bulk_dtend = _parse_iso_datetime(dtend) if dtend else None
+        if (dtstart and bulk_dtstart is None) or (dtend and bulk_dtend is None):
+            raise click.ClickException("--dtstart and --dtend must be ISO datetime strings (example: 2026-05-21T00:00:00Z).")
 
-    log_path = config_util.get('Paths', 'log')
-    if log_path == '':
-        log_path = os.path.join(os.path.dirname(abs_path), 'log',
-                               'logger.log')
-    elif not os.path.isabs(log_path):
-        log_path = os.path.join(os.path.dirname(abs_path),
-                               log_path)
-    config_params['log_path'] = log_path
+        payload = _safe_rapi_call(
+            rapi_api.list_order_items,
+            max_orders=limit,
+            dtstart=bulk_dtstart,
+            dtend=bulk_dtend,
+            status="AVAILABLE_FOR_DOWNLOAD",
+        )
+        filtered_items = list(payload or [])
 
-    # Set the timeout values
-    timeout_query = config_util.get('RAPI', 'timeout_query')
-    timeout_order = config_util.get('RAPI', 'timeout_order')
+    if not filtered_items:
+        click.echo("No AVAILABLE_FOR_DOWNLOAD order items found.")
+        return
 
-    try:
-        timeout_query = float(timeout_query)
-    except ValueError:
-        timeout_query = 60.0
+    click.echo(f"Found {len(filtered_items)} AVAILABLE_FOR_DOWNLOAD item(s).")
 
-    try:
-        timeout_order = float(timeout_order)
-    except ValueError:
-        timeout_order = 180.0
-    config_params['timeout_query'] = timeout_query
-    config_params['timeout_order'] = timeout_order
-
-    config_params['keep_results'] = config_util.get('Script', 'keep_results')
-    config_params['keep_downloads'] = config_util.get('Script',
-                                                      'keep_downloads')
-    config_params['colourize'] = config_util.get('Script', 'colourize')
-
-    # Get the total number of results per query
-    config_params['max_results'] = config_util.get('RAPI', 'max_results')
-
-    # Get the minimum date value to check orders
-    config_params['order_check_date'] = config_util.get('RAPI',
-                                                        'order_check_date')
-
-    config_params['download_attempts'] = config_util.get('RAPI',
-                                                        'download_attempts')
-    
-    config_params['concurrent_downloads'] = config_util.get('DDS',
-                                                        'concurrent_downloads')
-
-    return config_params
-
-def print_support(err_str=None):
-    """
-    Prints the 2 different support message depending if an error occurred.
-    
-    :param err_str: The error string to print along with support.
-    :type  err_str: str
-    """
-
-    eod_util.EodmsProcess().print_msg(err_str, heading='error')
-
-def get_latest_version():
-    package = 'py-eodms-rapi'  # replace with the package you want to check
-    response = requests.get(f'https://pypi.org/pypi/{package}/json')
-    latest_version = response.json()['info']['version']
-
-    return latest_version
-
-def setup_logger(log_name, log_path):
-
-    logger = logging.getLogger(log_name)
-
-    formatter = logging.Formatter('%(asctime)s - %(levelname)s - '
-                                    '%(message)s',
-                                    datefmt='%Y-%m-%d %I:%M:%S %p')
-    log_handler = handlers.RotatingFileHandler(log_path, maxBytes=500000,
-                                                backupCount=2)
-    log_handler.setLevel(logging.DEBUG)
-    log_handler.setFormatter(formatter)
-    logger.addHandler(log_handler)
-
-    return logger
-
-def print_choices():
-
-    output = 'The type of process to run from this list of options: '
-    for k, v in proc_choices.items():
-        output += f"'{k}': {v.get('desc')} | "
-
-    return output
-
-eodmsrapi_recent = get_latest_version()
-
-@click.command(context_settings={'help_option_names': ['-h', '--help']})
-@click.option('--configure', default=None,
-              help='Runs the configuration setup allowing the user to enter '
-                   'configuration values.')
-@click.option('--username', '-u', default=None,
-              help='The username of the EODMS account used for '
-                   'authentication.')
-@click.option('--password', '-p', default=None,
-              help='The password of the EODMS account used for '
-                   'authentication.')
-@click.option('--process', '-prc', '-r', default=None,
-              help=print_choices())
-@click.option('--input_val', '-i', default=None,
-              help='An input file (can either be an AOI, a CSV file '
-                   'exported from the EODMS UI), a WKT feature or a set '
-                   'of Record IDs. Valid AOI formats are GeoJSON, KML or '
-                   'Shapefile (Shapefile requires the GDAL Python '
-                   'package).')
-@click.option('--collections', default=None,
-              help='The collection of the images being ordered (separate '
-                   'multiple collections with a comma).')
-@click.option('--filters', '-f', default=None,
-              help='A list of filters for a specific collection.')
-@click.option('--dates', '-d', default=None,
-              help='The date ranges for the search.')
-@click.option('--maximum', '-max', '-m', default=None,
-              help='For Process 1, the maximum number of images to order '
-                   'and download. If no_order is set to True, this '
-                   'parameter will set the maximum images for which to search.')
-@click.option('--priority', '-pri', '-l', default=None,
-              help='The priority level of the order.\nOne of "Low", '
-                   '"Medium", "High" or "Urgent" (default "Medium").')
-@click.option('--output', '-o', default=None, help=output_help)
-@click.option('--aws', '-a', is_flag=True, default=None,
-              help='Determines whether to download from AWS (only applies '
-                   'to Radarsat-1 imagery).')
-@click.option('--overlap', '-ov', default=None,
-              help='The minimum percentage of overlap between AOI and images '
-                   '(if no AOI specified, this parameter is ignored).')
-@click.option('--orderitems', '-oid', default=None,
-              help="For Process 4, a set of Order IDs and/or Order Item IDs. "
-                   "This example specifies Order IDs and Order Item IDs: "
-                   "'order:151873,151872|item:1706113,1706111'")
-@click.option('--no_order', '-nord', is_flag=True, default=None,
-              help='If set, no ordering and downloading will occur.')
-@click.option('--downloads', '-dn', default=None,
-              help='The path where the images will be downloaded. Overrides '
-                   'the downloads parameter in the configuration file.')
-@click.option('--st_request', '-st', default=None,
-              help='The path of a file containing the JSON request for a ' 
-              'SAR Toolbox order.')
-@click.option('--silent', '-s', is_flag=True, default=None,
-              help='Sets process to silent which suppresses all questions.')
-@click.option('--version', '-v', is_flag=True, default=None,
-              help='Prints the version of the script.')
-@click.option('--uuid', default=None,
-              help='The UUID(s) of the items to order and download. Use with '
-                   '--collection for Process 3.')
-@click.option('--collection', '-c', default=None,
-              help='The collection ID for UUID-based ordering (Process 3). '
-                   'Use with --uuid.')
-def cli(username, password, input_val, collections, process, filters, dates,
-        maximum, priority, output, aws, overlap, orderitems, no_order,
-        downloads, st_request, silent, version, configure, uuid, collection):
-    """
-    Search & Order EODMS products.
-    """
-
-    os.system("title " + __title__)
-    sys.stdout.write("\x1b]2;%s\x07" % __title__)
-
-    python_version_cur = ".".join([str(sys.version_info.major),
-                                   str(sys.version_info.minor),
-                                   str(sys.version_info.micro)])
-    if pack_v.Version(python_version_cur) < pack_v.Version('3.6'):
-        raise Exception("Must be using Python 3.6 or higher")
-
-    if '-v' in sys.argv or '--v' in sys.argv or '--version' in sys.argv:
-        print(f"\n  {__title__}, version {__version__}\n")
-        eod_util.EodmsProcess().exit_cli()
-
-    conf_util = config_util.ConfigUtils(eod_util.EodmsProcess())
-
-    if configure:
-        conf_util.ask_user(configure)
-        eod_util.EodmsProcess().exit_cli()
-        
-    # Set all the parameters from the config.ini file
-    conf_util.import_config()
-
-    config_params = get_configuration_values(conf_util, downloads)
-    download_path = os.path.abspath(config_params['download_path'])
-    res_path = config_params['res_path']
-    log_path = config_params['log_path']
-    timeout_query = config_params['timeout_query']
-    timeout_order = config_params['timeout_order']
-    keep_results = config_params['keep_results']
-    keep_downloads = config_params['keep_downloads']
-    colourize = config_params['colourize']
-    max_results = config_params['max_results']
-    order_check_date = config_params['order_check_date']
-    download_attempts = config_params['download_attempts']
-    concurrent_downloads = config_params['concurrent_downloads']
+    _download_rapi_items(rapi_api, filtered_items, download_dir)
 
 
-    eodms_domain = os.getenv('EODMS_STAGING_DOMAIN')
-
-    print(eod_util.EodmsProcess(colourize=colourize).title_colour)
-    print("##########################################################"
-          "#######################")
-    print(f"#                              {__title__} v{__version__}         "
-          f"                        #")
-    print("############################################################"
-          "#####################")
-    print(eod_util.EodmsProcess(colourize=colourize).reset_colour)
-
-    rapi_installed_ver = eodms_rapi.__version__
-
-    path_colour = eod_util.EodmsProcess(colourize=colourize).path_colour
-    warn_colour = eod_util.EodmsProcess(colourize=colourize).warn_colour
-    if pack_v.Version(rapi_installed_ver) < pack_v.Version(min_rapi_version):
-        err_msg = f"The py-eodms-rapi currently installed is an older " \
-                    f"version (v{rapi_installed_ver}) than the minimum " \
-                    f"required version (v{min_rapi_version}). Please " \
-                    f"install it using: '{path_colour}pip install " \
-                    f"py-eodms-rapi -U{warn_colour}'."
-        eod_util.EodmsProcess().print_msg(err_msg, heading='error')
-        eod_util.EodmsProcess().exit_cli(1)
-
-    elif pack_v.Version(rapi_installed_ver) < \
-            pack_v.Version(eodmsrapi_recent):
-        msg = ""
-        msg = f"The py-eodms-rapi currently installed " \
-                f"(v{rapi_installed_ver}) is not the latest "\
-                f"version (v{eodmsrapi_recent}). It is recommended to use "\
-                f"the latest version of the package. Please install it " \
-                f"using: '{path_colour}pip install py-eodms-rapi -U" \
-                f"{warn_colour}'."
-        eod_util.EodmsProcess().print_msg(msg, heading="warning")
-
-    # Create info folder, if it doesn't exist, to store CSV files
-    start_time = datetime.datetime.now()
-    start_str = start_time.strftime("%Y-%m-%d %H:%M:%S")
-
-    eod = None
-    logger = None
-
-    try:
-
-        params = {'username': username,
-                  'password': password,
-                  'input_val': input_val,
-                  'collections': collections,
-                  'process': process,
-                  'filters': filters,
-                  'dates': dates,
-                  'maximum': maximum,
-                  'priority': priority,
-                  'output': output,
-                  # 'csv_fields': csv_fields,
-                  'aws': aws,
-                  'overlap': overlap,
-                  'orderitems': orderitems,
-                  'no_order': no_order,
-                  'downloads': downloads,
-                  'st_request': st_request,
-                  'silent': silent,
-                  'version': version,
-                  'uuid': uuid,
-                  'collection': collection}
-
-        fn_col = eod_util.EodmsProcess(colourize=colourize).path_colour
-        reset = eod_util.EodmsProcess(colourize=colourize).reset_colour
-        print(f"\nImages will be downloaded to " \
-            f"'{fn_col}{download_path}{reset}'.")
-
-        if not os.path.exists(os.path.dirname(log_path)):
-            pathlib.Path(os.path.dirname(log_path)).mkdir(
-                parents=True, exist_ok=True)
-            
-        # Setup logging for the RAPI Python package to print to .log file
-        logger = setup_logger('EODMSRAPI', log_path)
-
-        # Setup logging for the DDS Python package to print to .log file
-        setup_logger('eodms_dds', log_path)
-
-        logger.info(f"Script start time: {start_str}")
-
-        eod = eod_util.EodmsProcess(version=__version__, 
-                                    download=download_path,
-                                    results=res_path, log=log_path,
-                                    timeout_order=timeout_order,
-                                    timeout_query=timeout_query,
-                                    max_res=max_results,
-                                    keep_results=keep_results,
-                                    keep_downloads=keep_downloads,
-                                    colourize=colourize, 
-                                    order_check_date=order_check_date,
-                                    download_attempts=download_attempts,
-                                    eodms_domain=eodms_domain,
-                                    concurrent_downloads=concurrent_downloads)
-
-        eod.cleanup_folders()
-
-        #########################################
-        # Get authentication if not specified
-        #########################################
-
-        prmpt = Prompter(eod, conf_util, params)
-
-        prmpt.prompt()
-
-        eod.eodms_rapi.close_session()
-
-        prmpt.print_syntax()
-
-        eod.exit_cli()
-
-    except KeyboardInterrupt:
-        msg = "Process ended by user."
-        print(f"\n{msg}")
-
-        logger.info(msg)
-
-        if 'eod' in vars() or 'eod' in globals():
-            eod.exit_cli(1)
-        else:
-            eod_util.EodmsProcess().exit_cli(1)
-    except Exception:
-        trc_back = f"\n{traceback.format_exc()}"
-        logger.error(traceback.format_exc())
-        if 'eod' in vars() or 'eod' in globals():
-            eod.print_msg(trc_back, heading='error')
-            eod.exit_cli(0)
-        else:
-            eod_util.EodmsProcess().print_msg(trc_back, heading='error')
-            eod_util.EodmsProcess().exit_cli(0)
-
-if __name__ == '__main__':
+if __name__ == "__main__":
     cli()
